LWM2M: Split the module between different files
Since the LWM2M module has been increasing a lot latelly,
this patch splits the functionality of the module between
different, specific files, in order to make maintenance
easier.

Old file structure organization:
src/lib/comms/include/sol-lwm2m.h
src/lib/comms/sol-lwm2m.c

New file structure organization:
src/lib/comms/include/sol-lwm2m.h
src/lib/comms/include/sol-lwm2m-client.h
src/lib/comms/include/sol-lwm2m-server.h
src/lib/comms/include/sol-lwm2m-bs-server.h
src/lib/comms/sol-lwm2m-client.c
src/lib/comms/sol-lwm2m-server.c
src/lib/comms/sol-lwm2m-bs-server.c
src/lib/comms/sol-lwm2m-common.h
src/lib/comms/sol-lwm2m-common.c
