+++ src/lib/comms/sol-lwm2m-client.c
@@ -22,12 +22,14 @@
 #include <math.h>
 #include <ctype.h>
 
-#define SOL_LOG_DOMAIN &_lwm2m_domain
+#define SOL_LOG_DOMAIN &_lwm2m_client_domain
 
 #include "sol-log-internal.h"
 #include "sol-util-internal.h"
 #include "sol-list.h"
 #include "sol-lwm2m.h"
+#include "sol-lwm2m-common.h"
+#include "sol-lwm2m-client.h"
 #include "sol-macros.h"
 #include "sol-mainloop.h"
 #include "sol-monitors.h"
@@ -37,106 +39,7 @@
 #include "sol-util.h"
 #include "sol-http.h"
 
-SOL_LOG_INTERNAL_DECLARE_STATIC(_lwm2m_domain, "lwm2m");
-
-#define LWM2M_BOOTSTRAP_QUERY_PARAMS (1)
-#define LWM2M_UPDATE_QUERY_PARAMS (4)
-#define LWM2M_REGISTER_QUERY_PARAMS (5)
-#define NUMBER_OF_PATH_SEGMENTS (3)
-#define DEFAULT_SHORT_SERVER_ID (0)
-#define DEFAULT_CLIENT_LIFETIME (86400)
-#define DEFAULT_BINDING_MODE (SOL_LWM2M_BINDING_MODE_U)
-#define DEFAULT_LOCATION_PATH_SIZE (10)
-#define TLV_TYPE_MASK (192)
-#define TLV_ID_SIZE_MASK (32)
-#define TLV_CONTENT_LENGTH_MASK (24)
-#define TLV_CONTENT_LENGHT_CUSTOM_MASK (7)
-#define REMOVE_SIGN_BIT_MASK (127)
-#define SIGN_BIT_MASK (128)
-#define ID_HAS_16BITS_MASK (32)
-#define OBJ_LINK_LEN (4)
-#define LEN_IS_8BITS_MASK (8)
-#define LEN_IS_16BITS_MASK (16)
-#define LEN_IS_24BITS_MASK (24)
-#define UINT24_MAX (16777215)
-#define ONE_SECOND (1000)
-
-#define SECURITY_SERVER_OBJECT_ID (0)
-#define SECURITY_SERVER_URI (0)
-#define SECURITY_SERVER_IS_BOOTSTRAP (1)
-#define SECURITY_SERVER_ID (10)
-#define SECURITY_SERVER_CLIENT_HOLD_OFF_TIME (11)
-#define SECURITY_SERVER_BOOTSTRAP_SERVER_ACCOUNT_TIMEOUT (12)
-
-#define SERVER_OBJECT_ID (1)
-#define SERVER_OBJECT_SERVER_ID (0)
-#define SERVER_OBJECT_LIFETIME (1)
-#define SERVER_OBJECT_BINDING (7)
-
-#define ACCESS_CONTROL_OBJECT_ID (2)
-#define ACCESS_CONTROL_OBJECT_OBJECT_RES_ID (0)
-#define ACCESS_CONTROL_OBJECT_INSTANCE_RES_ID (1)
-#define ACCESS_CONTROL_OBJECT_ACL_RES_ID (2)
-#define ACCESS_CONTROL_OBJECT_OWNER_RES_ID (3)
-
-#ifndef SOL_NO_API_VERSION
-#define LWM2M_TLV_CHECK_API(_tlv, ...) \
-    do { \
-        if (SOL_UNLIKELY((_tlv)->api_version != \
-            SOL_LWM2M_TLV_API_VERSION)) { \
-            SOL_WRN("Couldn't handle tlv that has unsupported version " \
-                "'%u', expected version is '%u'", \
-                (_tlv)->api_version, SOL_LWM2M_TLV_API_VERSION); \
-            return __VA_ARGS__; \
-        } \
-    } while (0);
-#define LWM2M_RESOURCE_CHECK_API(_resource, ...) \
-    do { \
-        if (SOL_UNLIKELY((_resource)->api_version != \
-            SOL_LWM2M_RESOURCE_API_VERSION)) { \
-            SOL_WRN("Couldn't handle resource that has unsupported version " \
-                "'%u', expected version is '%u'", \
-                (_resource)->api_version, SOL_LWM2M_RESOURCE_API_VERSION); \
-            return __VA_ARGS__; \
-        } \
-    } while (0);
-#define LWM2M_RESOURCE_CHECK_API_GOTO(_resource, _label) \
-    do { \
-        if (SOL_UNLIKELY((_resource).api_version != \
-            SOL_LWM2M_RESOURCE_API_VERSION)) { \
-            SOL_WRN("Couldn't handle resource that has unsupported version " \
-                "'%u', expected version is '%u'", \
-                (_resource).api_version, SOL_LWM2M_RESOURCE_API_VERSION); \
-            goto _label; \
-        } \
-    } while (0);
-#define LWM2M_OBJECT_CHECK_API(_obj, ...) \
-    do { \
-        if (SOL_UNLIKELY((_obj).api_version != \
-            SOL_LWM2M_OBJECT_API_VERSION)) { \
-            SOL_WRN("Couldn't handle object that has unsupported version " \
-                "'%u', expected version is '%u'", \
-                (_obj).api_version, SOL_LWM2M_OBJECT_API_VERSION); \
-            return __VA_ARGS__; \
-        } \
-    } while (0);
-#define LWM2M_OBJECT_CHECK_API_GOTO(_obj, _label) \
-    do { \
-        if (SOL_UNLIKELY((_obj).api_version != \
-            SOL_LWM2M_OBJECT_API_VERSION)) { \
-            SOL_WRN("Couldn't handle object that has unsupported version " \
-                "'%u', expected version is '%u'", \
-                (_obj).api_version, SOL_LWM2M_OBJECT_API_VERSION); \
-            goto _label; \
-        } \
-    } while (0);
-#else
-#define LWM2M_TLV_CHECK_API(_tlv, ...)
-#define LWM2M_RESOURCE_CHECK_API(_resource, ...)
-#define LWM2M_OBJECT_CHECK_API(_obj, ...)
-#define LWM2M_RESOURCE_CHECK_API_GOTO(_resource, _label)
-#define LWM2M_OBJECT_CHECK_API_GOTO(_obj, _label)
-#endif
+SOL_LOG_INTERNAL_DECLARE_STATIC(_lwm2m_client_domain, "lwm2m-client");
 
 #define ADD_QUERY(_key, _format, _value) \
     do { \
@@ -147,106 +50,6 @@ SOL_LOG_INTERNAL_DECLARE_STATIC(_lwm2m_domain, "lwm2m");
             query.data, query.used); \
         SOL_INT_CHECK_GOTO(r, < 0, err_coap); \
     } while (0);
-
-enum tlv_length_size_type {
-    LENGTH_SIZE_CHECK_NEXT_TWO_BITS = 0,
-    LENGTH_SIZE_8_BITS = 8,
-    LENGTH_SIZE_16_BITS = 16,
-    LENGTH_SIZE_24_BITS = 24
-};
-
-enum lwm2m_parser_args_state {
-    STATE_NEEDS_DIGIT = 0,
-    STATE_NEEDS_COMMA_OR_EQUAL = (1 << 1),
-    STATE_NEEDS_COMMA = (1 << 2),
-    STATE_NEEDS_APOSTROPHE = (1 << 3),
-    STATE_NEEDS_CHAR_OR_DIGIT = (1 << 4),
-};
-
-struct lifetime_ctx {
-    struct sol_timeout *timeout;
-    uint32_t lifetime;
-};
-
-struct sol_lwm2m_server {
-    struct sol_coap_server *coap;
-    struct sol_ptr_vector clients;
-    struct sol_ptr_vector clients_to_delete;
-    struct sol_monitors registration;
-    struct sol_ptr_vector observers;
-    struct lifetime_ctx lifetime_ctx;
-};
-
-struct sol_lwm2m_bootstrap_server {
-    struct sol_coap_server *coap;
-    struct sol_ptr_vector clients;
-    struct sol_monitors bootstrap;
-    const char **known_clients;
-};
-
-struct sol_lwm2m_client_object {
-    struct sol_ptr_vector instances;
-    uint16_t id;
-};
-
-struct sol_lwm2m_client_info {
-    struct sol_ptr_vector objects;
-    char *name;
-    char *location;
-    char *sms;
-    char *objects_path;
-    uint32_t lifetime;
-    time_t register_time;
-    struct sol_lwm2m_server *server;
-    struct sol_network_link_addr cliaddr;
-    enum sol_lwm2m_binding_mode binding;
-    struct sol_coap_resource resource;
-};
-
-struct sol_lwm2m_bootstrap_client_info {
-    char *name;
-    struct sol_network_link_addr cliaddr;
-};
-
-struct observer_entry {
-    struct sol_monitors monitors;
-    struct sol_lwm2m_server *server;
-    struct sol_lwm2m_client_info *cinfo;
-    int64_t token;
-    char *path;
-    bool removed;
-};
-
-enum management_type {
-    MANAGEMENT_DELETE,
-    MANAGEMENT_READ,
-    MANAGEMENT_CREATE,
-    MANAGEMENT_WRITE,
-    MANAGEMENT_EXECUTE
-};
-
-struct management_ctx {
-    enum management_type type;
-    struct sol_lwm2m_server *server;
-    struct sol_lwm2m_client_info *cinfo;
-    char *path;
-    void *cb;
-    const void *data;
-};
-
-enum bootstrap_type {
-    BOOTSTRAP_DELETE,
-    BOOTSTRAP_WRITE
-};
-
-struct bootstrap_ctx {
-    enum bootstrap_type type;
-    struct sol_lwm2m_bootstrap_server *server;
-    struct sol_lwm2m_bootstrap_client_info *cinfo;
-    char *path;
-    void *cb;
-    const void *data;
-};
 
 struct notification_ctx {
     struct sol_lwm2m_client *client;
@@ -261,7 +64,6 @@ struct resource_ctx {
     uint16_t id;
 };
 
-//Data structs used by LWM2M Client.
 struct obj_instance {
     uint16_t id;
     bool should_delete;
@@ -313,7 +115,6 @@ struct server_conn_ctx {
     char *location;
 };
 
-static bool lifetime_server_timeout(void *data);
 static bool lifetime_client_timeout(void *data);
 static int register_with_server(struct sol_lwm2m_client *client,
     struct server_conn_ctx *conn_ctx, bool is_update);
@@ -332,40 +133,6 @@ static int
 read_resources(struct sol_lwm2m_client *client,
     struct obj_ctx *obj_ctx, struct obj_instance *instance,
     struct sol_lwm2m_resource *res, size_t res_len, ...);
-static void
-clear_resource_array(struct sol_lwm2m_resource *array, uint16_t len);
-
-static void
-send_ack_if_needed(struct sol_coap_server *coap, struct sol_coap_packet *msg,
-    const struct sol_network_link_addr *cliaddr)
-{
-    struct sol_coap_packet *ack;
-    uint8_t type;
-
-    sol_coap_header_get_type(msg, &type);
-
-    if (type == SOL_COAP_MESSAGE_TYPE_CON) {
-        ack = sol_coap_packet_new(msg);
-        SOL_NULL_CHECK(ack);
-        if (sol_coap_send_packet(coap, ack, cliaddr) < 0)
-            SOL_WRN("Could not send the response ACK");
-    }
-}
-
-static void
-dispatch_registration_event(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *cinfo,
-    enum sol_lwm2m_registration_event event)
-{
-    uint16_t i;
-    struct sol_monitors_entry *m;
-
-    SOL_MONITORS_WALK (&server->registration, m, i)
-        ((void (*)(void *, struct sol_lwm2m_server *,
-        struct sol_lwm2m_client_info *,
-        enum sol_lwm2m_registration_event))m->cb)((void *)m->data, server,
-            cinfo, event);
-}
 
 static void
 dispatch_bootstrap_event_to_client(struct sol_lwm2m_client *client,
@@ -377,1098 +144,6 @@ dispatch_bootstrap_event_to_client(struct sol_lwm2m_client *client,
     SOL_MONITORS_WALK (&client->bootstrap, m, i)
         ((void (*)(void *, struct sol_lwm2m_client *,
         enum sol_lwm2m_bootstrap_event))m->cb)((void *)m->data, client, event);
-}
-
-static void
-dispatch_bootstrap_event_to_server(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *bs_cinfo)
-{
-    uint16_t i;
-    struct sol_monitors_entry *m;
-
-    SOL_MONITORS_WALK (&server->bootstrap, m, i)
-        ((void (*)(void *, struct sol_lwm2m_bootstrap_server *,
-        struct sol_lwm2m_bootstrap_client_info *))m->cb)((void *)m->data, server,
-            bs_cinfo);
-}
-
-static void
-client_objects_clear(struct sol_ptr_vector *objects)
-{
-    uint16_t i, j, *id;
-    struct sol_lwm2m_client_object *object;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (objects, object, i) {
-        SOL_PTR_VECTOR_FOREACH_IDX (&object->instances, id, j)
-            free(id);
-        sol_ptr_vector_clear(&object->instances);
-        free(object);
-    }
-
-    sol_ptr_vector_clear(objects);
-}
-
-static void
-bootstrap_client_info_del(struct sol_lwm2m_bootstrap_client_info *bs_cinfo)
-{
-    free(bs_cinfo->name);
-    free(bs_cinfo);
-}
-
-static void
-client_info_del(struct sol_lwm2m_client_info *cinfo)
-{
-    free(cinfo->sms);
-    free(cinfo->location);
-    free(cinfo->name);
-    free(cinfo->objects_path);
-    client_objects_clear(&cinfo->objects);
-    free(cinfo);
-}
-
-static enum sol_lwm2m_binding_mode
-get_binding_mode_from_str(const struct sol_str_slice binding)
-{
-    static const struct sol_str_table map[] = {
-        SOL_STR_TABLE_PTR_ITEM("U", SOL_LWM2M_BINDING_MODE_U),
-        //TODO: The modes below are not supported for now.
-        SOL_STR_TABLE_PTR_ITEM("UQ", SOL_LWM2M_BINDING_MODE_UNKNOWN),
-        SOL_STR_TABLE_PTR_ITEM("S", SOL_LWM2M_BINDING_MODE_UNKNOWN),
-        SOL_STR_TABLE_PTR_ITEM("SQ", SOL_LWM2M_BINDING_MODE_UNKNOWN),
-        SOL_STR_TABLE_PTR_ITEM("US", SOL_LWM2M_BINDING_MODE_UNKNOWN),
-        SOL_STR_TABLE_PTR_ITEM("UQS", SOL_LWM2M_BINDING_MODE_UNKNOWN),
-        { }
-    };
-
-    return sol_str_table_lookup_fallback(map, binding,
-        SOL_LWM2M_BINDING_MODE_UNKNOWN);
-}
-
-static void
-clients_to_delete_clear(struct sol_ptr_vector *to_delete)
-{
-    uint16_t i;
-    struct sol_lwm2m_client_info *cinfo;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (to_delete, cinfo, i)
-        client_info_del(cinfo);
-    sol_ptr_vector_clear(to_delete);
-}
-
-static void
-remove_all_observer_entries_from_client(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *cinfo)
-{
-    uint16_t i;
-    int64_t token;
-    struct observer_entry *entry;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->observers, entry, i) {
-        if (entry->cinfo == cinfo) {
-            token = entry->token;
-            entry->removed = true;
-            sol_coap_unobserve_by_token(server->coap, &cinfo->cliaddr,
-                (uint8_t *)&token, sizeof(token));
-        }
-    }
-}
-
-static void
-remove_client(struct sol_lwm2m_client_info *cinfo, bool del)
-{
-    int r = 0;
-
-    remove_all_observer_entries_from_client(cinfo->server, cinfo);
-
-    r = sol_ptr_vector_remove(&cinfo->server->clients, cinfo);
-    if (r < 0)
-        SOL_WRN("Could not remove the client %s from the clients list",
-            cinfo->name);
-    r = sol_coap_server_unregister_resource(cinfo->server->coap,
-        &cinfo->resource);
-    if (r < 0)
-        SOL_WRN("Could not unregister coap resource for the client: %s",
-            cinfo->name);
-    if (del)
-        client_info_del(cinfo);
-    else {
-        r = sol_ptr_vector_append(&cinfo->server->clients_to_delete, cinfo);
-        if (r < 0)
-            SOL_WRN("Could not add the client to pending clients list");
-    }
-}
-
-static struct sol_lwm2m_client_object *
-find_client_object_by_id(struct sol_ptr_vector *objects,
-    uint16_t id)
-{
-    uint16_t i;
-    struct sol_lwm2m_client_object *cobject;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (objects, cobject, i) {
-        if (cobject->id == id)
-            return cobject;
-    }
-
-    return NULL;
-}
-
-static int
-fill_client_objects(struct sol_lwm2m_client_info *cinfo,
-    struct sol_coap_packet *req, bool update)
-{
-    struct sol_lwm2m_client_object *cobject;
-    struct sol_str_slice content, *object;
-    struct sol_vector objects;
-    struct sol_buffer *buf;
-    uint16_t *instance;
-    bool has_content;
-    size_t offset;
-    uint16_t i;
-    int r;
-
-#define TO_INT(_data, _endptr, _len, _i, _label) \
-    _i = sol_util_strtol_n(_data, &_endptr, _len, 10); \
-    if (_endptr == _data || errno != 0 ) { \
-        SOL_WRN("Could not convert object to int. (%.*s)", \
-            SOL_STR_SLICE_PRINT(*object)); \
-        r = -EINVAL; \
-        goto _label; \
-    }
-#define EXIT_IF_FAIL(_condition) \
-    if (_condition) { \
-        r = -EINVAL; \
-        SOL_WRN("Malformed object: %.*s", \
-            SOL_STR_SLICE_PRINT(*object)); \
-        goto err_exit; \
-    }
-
-    has_content = sol_coap_packet_has_payload(req);
-
-    if (!has_content && !update) {
-        SOL_WRN("The registration request has no payload!");
-        return -ENOENT;
-    } else if (!has_content)
-        return 0;
-
-    client_objects_clear(&cinfo->objects);
-
-    r = sol_coap_packet_get_payload(req, &buf, &offset);
-    SOL_INT_CHECK(r, < 0, r);
-    content.data = sol_buffer_at(buf, offset);
-    content.len = buf->used - offset;
-
-    SOL_DBG("Register payload content: %.*s", (int)content.len, content.data);
-    objects = sol_str_slice_split(content, ",", 0);
-
-    if (!objects.len) {
-        SOL_WRN("The objects list is empty!");
-        return -EINVAL;
-    }
-
-    SOL_VECTOR_FOREACH_IDX (&objects, object, i) {
-        char *endptr;
-        uint16_t id;
-
-        *object = sol_str_slice_trim(*object);
-
-        EXIT_IF_FAIL(object->len < 4 || object->data[0] != '<');
-
-        /*
-           Object form: </ObjectId[/InstanceID]>
-           Where ObjectId is an integer (must be present)
-           InstanceId is an integer, may not be present and can not be UINT16_MAX
-           Alternate path: </a/path>[;rt="oma.lwm2m"][;ct=1058]
-         */
-        if (sol_str_slice_str_contains(*object, "rt=\"oma.lwm2m\"")) {
-            struct sol_str_slice path;
-            endptr = memrchr(object->data, '>', object->len);
-            EXIT_IF_FAIL(!endptr);
-            path.data = object->data + 1;
-            path.len = endptr - path.data;
-            r = sol_util_replace_str_from_slice_if_changed(&cinfo->objects_path,
-                path);
-            SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-            if (cinfo->objects_path && streq(cinfo->objects_path, "/")) {
-                free(cinfo->objects_path);
-                cinfo->objects_path = NULL;
-            }
-            continue;
-        } else if (sol_str_slice_str_contains(*object, "ct=")) {
-            //The content type value for json was not defined yet.
-            //TODO: Support JSON formats.
-            SOL_WRN("Only text format is supported for now");
-            r = -EINVAL;
-            goto err_exit;
-        }
-
-        EXIT_IF_FAIL(object->data[object->len - 1] != '>');
-
-        //Removing '<', '>' and '/'
-        object->data += 2;
-        object->len -= 3;
-
-        TO_INT(object->data, endptr, object->len, id, err_exit);
-
-        cobject = find_client_object_by_id(&cinfo->objects, id);
-
-        if (!cobject) {
-            cobject = malloc(sizeof(struct sol_lwm2m_client_object));
-            r = -ENOMEM;
-            SOL_NULL_CHECK_GOTO(cobject, err_exit);
-            r = sol_ptr_vector_append(&cinfo->objects, cobject);
-            if (r < 0) {
-                SOL_WRN("Could not append the object id:%" PRIu16
-                    " to the object array", id);
-                free(cobject);
-                goto err_exit;
-            }
-            sol_ptr_vector_init(&cobject->instances);
-            cobject->id = id;
-        }
-
-        //Advance to instance ID
-        object->len -= endptr - object->data;
-
-        //Instance ID not provided.
-        if (!object->len)
-            continue;
-
-        //Skip '/'
-        object->data = endptr + 1;
-        object->len--;
-
-        instance = malloc(sizeof(uint16_t));
-        r = -ENOMEM;
-        SOL_NULL_CHECK_GOTO(instance, err_exit);
-
-        TO_INT(object->data, endptr, object->len, *instance,
-            err_instance_to_int);
-
-        if (*instance == UINT16_MAX) {
-            SOL_WRN("The instance id value: %" PRIu16 " must not be used!",
-                UINT16_MAX);
-            r = -EPERM;
-            goto err_instance_to_int;
-        }
-
-        r = sol_ptr_vector_append(&cobject->instances, instance);
-        if (r < 0) {
-            SOL_WRN("Could not append the instance /%" PRIu16 "/%" PRIu16
-                " to the instance array", cobject->id, *instance);
-            goto err_instance_to_int;
-        }
-    }
-
-    sol_vector_clear(&objects);
-    return 0;
-
-err_instance_to_int:
-    free(instance);
-err_exit:
-    sol_vector_clear(&objects);
-    client_objects_clear(&cinfo->objects);
-    return r;
-
-#undef EXIT_IF_FAIL
-#undef TO_INT
-}
-
-static int
-fill_client_info(struct sol_lwm2m_client_info *cinfo,
-    struct sol_coap_packet *req, bool update)
-{
-    uint16_t i, count;
-    bool has_name = false;
-    struct sol_str_slice query[5];
-    int r;
-
-    r = sol_coap_find_options(req, SOL_COAP_OPTION_URI_QUERY, query,
-        update ? LWM2M_UPDATE_QUERY_PARAMS : LWM2M_REGISTER_QUERY_PARAMS);
-    SOL_INT_CHECK(r, < 0, r);
-    count = r;
-    cinfo->register_time = time(NULL);
-
-    for (i = 0; i < count; i++) {
-        struct sol_str_slice key, value;
-        const char *sep;
-
-        SOL_DBG("Query:%.*s", SOL_STR_SLICE_PRINT(query[i]));
-        sep = memchr(query[i].data, '=', query[i].len);
-
-        if (!sep) {
-            SOL_WRN("Could not find the separator '=' at: %.*s",
-                SOL_STR_SLICE_PRINT(query[i]));
-            break;
-        }
-
-        key.data = query[i].data;
-        key.len = sep - query[i].data;
-        value.data = sep + 1;
-        value.len = query[i].len - key.len - 1;
-
-        if (sol_str_slice_str_eq(key, "ep")) {
-            if (update) {
-                SOL_WRN("The lwm2m client can not update it's name"
-                    " during the update");
-                r = -EPERM;
-                goto err_cinfo_prop;
-            }
-            //Required info
-            has_name = true;
-            cinfo->name = sol_str_slice_to_str(value);
-            SOL_NULL_CHECK_GOTO(cinfo->name, err_cinfo_prop);
-        } else if (sol_str_slice_str_eq(key, "lt")) {
-            char *endptr;
-            cinfo->lifetime = sol_util_strtoul_n(value.data, &endptr,
-                value.len, 10);
-            if (endptr == value.data || errno != 0) {
-                SOL_WRN("Could not convert the lifetime to integer."
-                    " Lifetime: %.*s", SOL_STR_SLICE_PRINT(value));
-                r = -EINVAL;
-                goto err_cinfo_prop;
-            }
-        } else if (sol_str_slice_str_eq(key, "sms")) {
-            r = sol_util_replace_str_from_slice_if_changed(&cinfo->sms, value);
-            SOL_INT_CHECK_GOTO(r, < 0, err_cinfo_prop);
-        } else if (sol_str_slice_str_eq(key, "lwm2m") &&
-            !sol_str_slice_str_eq(value, "1.0")) {
-            r = -EINVAL;
-            SOL_WRN("LWM2M version not supported:%.*s",
-                SOL_STR_SLICE_PRINT(value));
-            goto err_cinfo_prop;
-        } else if (sol_str_slice_str_eq(key, "b")) {
-            cinfo->binding = get_binding_mode_from_str(value);
-            r = -EINVAL;
-            SOL_INT_CHECK_GOTO(cinfo->binding,
-                == SOL_LWM2M_BINDING_MODE_UNKNOWN, err_cinfo_prop);
-        }
-    }
-
-    if (has_name || update)
-        return fill_client_objects(cinfo, req, update);
-    else {
-        SOL_WRN("The client did not provide its name!");
-        return -EINVAL;
-    }
-
-err_cinfo_prop:
-    return r;
-}
-
-static int
-extract_bootstrap_client_info(struct sol_coap_packet *req,
-    struct sol_str_slice *client_name)
-{
-    struct sol_str_slice query;
-    int r;
-
-    r = sol_coap_find_options(req, SOL_COAP_OPTION_URI_QUERY, &query,
-        LWM2M_BOOTSTRAP_QUERY_PARAMS);
-    SOL_INT_CHECK(r, < 0, r);
-
-    struct sol_str_slice key, value;
-    const char *sep;
-
-    SOL_DBG("Query:%.*s", SOL_STR_SLICE_PRINT(query));
-    sep = memchr(query.data, '=', query.len);
-
-    if (!sep) {
-        SOL_WRN("Could not find the separator '=' at: %.*s",
-            SOL_STR_SLICE_PRINT(query));
-        return -EINVAL;
-    }
-
-    key.data = query.data;
-    key.len = sep - query.data;
-    value.data = sep + 1;
-    value.len = query.len - key.len - 1;
-
-    if (sol_str_slice_str_eq(key, "ep")) {
-        //Required info
-        *client_name = value;
-    } else {
-        SOL_WRN("The client did not provide its name!");
-        return -EINVAL;
-    }
-
-    return 0;
-}
-
-static int
-reschedule_timeout(struct sol_lwm2m_server *server)
-{
-    struct sol_lwm2m_client_info *cinfo;
-    uint32_t smallest_remaining, remaining, lf = 0;
-    time_t now;
-    uint16_t i;
-    int r;
-
-    clients_to_delete_clear(&server->clients_to_delete);
-
-    if (server->lifetime_ctx.timeout)
-        sol_timeout_del(server->lifetime_ctx.timeout);
-
-    if (!sol_ptr_vector_get_len(&server->clients)) {
-        server->lifetime_ctx.timeout = NULL;
-        server->lifetime_ctx.lifetime = 0;
-        return 0;
-    }
-
-    smallest_remaining = UINT32_MAX;
-    now = time(NULL);
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->clients, cinfo, i) {
-        remaining = cinfo->lifetime - (now - cinfo->register_time);
-        if (remaining < smallest_remaining) {
-            smallest_remaining = remaining;
-            lf = cinfo->lifetime;
-        }
-    }
-
-    //Set to NULL in case we fail.
-    server->lifetime_ctx.timeout = NULL;
-    /*
-       When a client is registered, it tells the server what is its lifetime.
-       If the server's timeout is registered using the exactly same amount,
-       there's a high chance that the server will end up removing a client from
-       my list, because the message will take some time until it arrives
-       from the network. In order to reduce the change from happening,
-       the server will add 2 seconds to smallest_remaining.
-     */
-    r = sol_util_uint32_mul(smallest_remaining + 2, 1000, &smallest_remaining);
-    SOL_INT_CHECK(r, < 0, r);
-    server->lifetime_ctx.timeout = sol_timeout_add(smallest_remaining,
-        lifetime_server_timeout, server);
-    SOL_NULL_CHECK(server->lifetime_ctx.timeout, -ENOMEM);
-    server->lifetime_ctx.lifetime = lf;
-    return 0;
-}
-
-static bool
-lifetime_server_timeout(void *data)
-{
-    struct sol_ptr_vector to_delete = SOL_PTR_VECTOR_INIT;
-    struct sol_lwm2m_server *server = data;
-    struct sol_lwm2m_client_info *cinfo;
-    uint16_t i;
-    int r;
-
-    SOL_DBG("Lifetime timeout! (%" PRIu32 ")", server->lifetime_ctx.lifetime);
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->clients, cinfo, i) {
-        if (server->lifetime_ctx.lifetime != cinfo->lifetime)
-            continue;
-        SOL_DBG("Deleting client %s for inactivity", cinfo->name);
-        r = sol_ptr_vector_append(&to_delete, cinfo);
-        SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-    }
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&to_delete, cinfo, i) {
-        dispatch_registration_event(server, cinfo,
-            SOL_LWM2M_REGISTRATION_EVENT_TIMEOUT);
-        remove_client(cinfo, true);
-    }
-
-    sol_ptr_vector_clear(&to_delete);
-
-    r = reschedule_timeout(server);
-    if (r < 0)
-        SOL_WRN("Could not reschedule the lifetime timeout");
-    return false;
-
-err_exit:
-    sol_ptr_vector_clear(&to_delete);
-    return true;
-}
-
-static int
-update_client(void *data, struct sol_coap_server *coap,
-    const struct sol_coap_resource *resource,
-    struct sol_coap_packet *req,
-    const struct sol_network_link_addr *cliaddr)
-{
-    struct sol_lwm2m_client_info *cinfo = data;
-    struct sol_coap_packet *response;
-    int r;
-
-    SOL_DBG("Client update request (name: %s)", cinfo->name);
-
-    response = sol_coap_packet_new(req);
-    SOL_NULL_CHECK(response, -ENOMEM);
-
-    r = fill_client_info(cinfo, req, true);
-    SOL_INT_CHECK_GOTO(r, < 0, err_update);
-
-    r = reschedule_timeout(cinfo->server);
-    SOL_INT_CHECK_GOTO(r, < 0, err_update);
-
-    dispatch_registration_event(cinfo->server, cinfo,
-        SOL_LWM2M_REGISTRATION_EVENT_UPDATE);
-
-    r = sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_CHANGED);
-    SOL_INT_CHECK_GOTO(r, < 0, err_update);
-    return sol_coap_send_packet(coap, response, cliaddr);
-
-err_update:
-    sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_BAD_REQUEST);
-    sol_coap_send_packet(coap, response, cliaddr);
-    return r;
-}
-
-static int
-delete_client(void *data, struct sol_coap_server *coap,
-    const struct sol_coap_resource *resource,
-    struct sol_coap_packet *req,
-    const struct sol_network_link_addr *cliaddr)
-{
-    struct sol_lwm2m_client_info *cinfo = data;
-    struct sol_coap_packet *response;
-    int r;
-
-    SOL_DBG("Client delete request (name: %s)", cinfo->name);
-
-    response = sol_coap_packet_new(req);
-    SOL_NULL_CHECK(response, -ENOMEM);
-
-    remove_client(cinfo, false);
-
-    if (!sol_ptr_vector_get_len(&cinfo->server->clients) &&
-        cinfo->server->lifetime_ctx.timeout) {
-        sol_timeout_del(cinfo->server->lifetime_ctx.timeout);
-        cinfo->server->lifetime_ctx.timeout = NULL;
-        cinfo->server->lifetime_ctx.lifetime = 0;
-        SOL_DBG("Client list is empty");
-    }
-
-    dispatch_registration_event(cinfo->server, cinfo,
-        SOL_LWM2M_REGISTRATION_EVENT_UNREGISTER);
-
-    r = sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_DELETED);
-    SOL_INT_CHECK_GOTO(r, < 0, err);
-    return sol_coap_send_packet(coap, response, cliaddr);
-
-err:
-    sol_coap_packet_unref(response);
-    return r;
-}
-
-static int
-generate_location(char **location)
-{
-    SOL_BUFFER_DECLARE_STATIC(uuid, 33);
-    int r;
-
-    r = sol_util_uuid_gen(false, false, &uuid);
-    SOL_INT_CHECK(r, < 0, r);
-    *location = strndup(uuid.data, DEFAULT_LOCATION_PATH_SIZE);
-    SOL_NULL_CHECK(*location, -ENOMEM);
-    return 0;
-}
-
-static int
-new_client_info(struct sol_lwm2m_client_info **cinfo,
-    const struct sol_network_link_addr *cliaddr,
-    struct sol_lwm2m_server *server)
-{
-    int r;
-
-    *cinfo = calloc(1, sizeof(struct sol_lwm2m_client_info) +
-        (sizeof(struct sol_str_slice) * NUMBER_OF_PATH_SEGMENTS));
-    SOL_NULL_CHECK(cinfo, -ENOMEM);
-
-    (*cinfo)->lifetime = DEFAULT_CLIENT_LIFETIME;
-    (*cinfo)->binding = DEFAULT_BINDING_MODE;
-    r = generate_location(&(*cinfo)->location);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-    (*cinfo)->resource.flags = SOL_COAP_FLAGS_NONE;
-    (*cinfo)->resource.path[0] = sol_str_slice_from_str("rd");
-    (*cinfo)->resource.path[1] = sol_str_slice_from_str((*cinfo)->location);
-    (*cinfo)->resource.path[2] = sol_str_slice_from_str("");
-    (*cinfo)->resource.del = delete_client;
-    /*
-       Current spec says that the client update should be handled using
-       the post method, however some old clients still uses put.
-     */
-    (*cinfo)->resource.post = update_client;
-    (*cinfo)->resource.put = update_client;
-    (*cinfo)->server = server;
-    sol_ptr_vector_init(&(*cinfo)->objects);
-    memcpy(&(*cinfo)->cliaddr, cliaddr, sizeof(struct sol_network_link_addr));
-    SOL_SET_API_VERSION((*cinfo)->resource.api_version = SOL_COAP_RESOURCE_API_VERSION; )
-    return 0;
-err_exit:
-    free(*cinfo);
-    return r;
-}
-
-static int
-new_bootstrap_client_info(struct sol_lwm2m_bootstrap_client_info **bs_cinfo,
-    const struct sol_network_link_addr *cliaddr, const struct sol_str_slice client_name)
-{
-    *bs_cinfo = calloc(1, sizeof(struct sol_lwm2m_bootstrap_client_info));
-    SOL_NULL_CHECK(bs_cinfo, -ENOMEM);
-
-    memcpy(&(*bs_cinfo)->cliaddr, cliaddr, sizeof(struct sol_network_link_addr));
-
-    (*bs_cinfo)->name = sol_str_slice_to_str(client_name);
-    SOL_NULL_CHECK_GOTO((*bs_cinfo)->name, err_no_name);
-    return 0;
-
-err_no_name:
-    free(*bs_cinfo);
-    return -ENOMEM;
-}
-
-static struct sol_lwm2m_client_info *
-get_client_info_by_name(struct sol_ptr_vector *clients,
-    const char *name)
-{
-    struct sol_lwm2m_client_info *cinfo;
-    uint16_t i;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (clients, cinfo, i) {
-        if (streq(name, cinfo->name))
-            return cinfo;
-    }
-
-    return NULL;
-}
-
-static int
-registration_request(void *data, struct sol_coap_server *coap,
-    const struct sol_coap_resource *resource,
-    struct sol_coap_packet *req,
-    const struct sol_network_link_addr *cliaddr)
-{
-    struct sol_lwm2m_client_info *cinfo, *old_cinfo;
-    struct sol_lwm2m_server *server = data;
-    struct sol_coap_packet *response;
-    int r;
-
-    SOL_DBG("Client registration request");
-
-    response = sol_coap_packet_new(req);
-    SOL_NULL_CHECK(response, -ENOMEM);
-
-    r = new_client_info(&cinfo, cliaddr, server);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-    r = fill_client_info(cinfo, req, false);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_del_client);
-
-    old_cinfo = get_client_info_by_name(&server->clients,
-        cinfo->name);
-    if (old_cinfo) {
-        SOL_DBG("Client %s already exists, replacing it.", old_cinfo->name);
-        remove_client(old_cinfo, true);
-    }
-
-    r = sol_coap_server_register_resource(server->coap, &cinfo->resource,
-        cinfo);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_del_client);
-
-    r = sol_ptr_vector_append(&server->clients, cinfo);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_unregister);
-
-    r = reschedule_timeout(server);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_unregister);
-
-    r = sol_coap_add_option(response, SOL_COAP_OPTION_LOCATION_PATH,
-        "rd", strlen("rd"));
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_unregister);
-    r = sol_coap_add_option(response,
-        SOL_COAP_OPTION_LOCATION_PATH, cinfo->location, strlen(cinfo->location));
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_unregister);
-
-    r = sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_CREATED);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_unregister);
-
-    SOL_DBG("Client %s registered. Location: %s, SMS: %s, binding: %u,"
-        " lifetime: %" PRIu32 " objects paths: %s",
-        cinfo->name, cinfo->location, cinfo->sms,
-        cinfo->binding, cinfo->lifetime, cinfo->objects_path);
-
-    r = sol_coap_send_packet(coap, response, cliaddr);
-    dispatch_registration_event(server, cinfo,
-        SOL_LWM2M_REGISTRATION_EVENT_REGISTER);
-    return r;
-
-err_exit_unregister:
-    if (sol_coap_server_unregister_resource(server->coap, &cinfo->resource) < 0)
-        SOL_WRN("Could not unregister resource for client: %s", cinfo->name);
-err_exit_del_client:
-    client_info_del(cinfo);
-err_exit:
-    sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_BAD_REQUEST);
-    sol_coap_send_packet(coap, response, cliaddr);
-    return r;
-}
-
-static const struct sol_coap_resource registration_interface = {
-    SOL_SET_API_VERSION(.api_version = SOL_COAP_RESOURCE_API_VERSION, )
-    .post = registration_request,
-    .flags = SOL_COAP_FLAGS_NONE,
-    .path = {
-        SOL_STR_SLICE_LITERAL("rd"),
-        SOL_STR_SLICE_EMPTY
-    }
-};
-
-static int
-bootstrap_request(void *data, struct sol_coap_server *coap,
-    const struct sol_coap_resource *resource,
-    struct sol_coap_packet *req,
-    const struct sol_network_link_addr *cliaddr)
-{
-    struct sol_lwm2m_bootstrap_client_info *bs_cinfo;
-    struct sol_lwm2m_bootstrap_server *server = data;
-    struct sol_coap_packet *response;
-    struct sol_str_slice client_name = SOL_STR_SLICE_EMPTY;
-    int r;
-    size_t i;
-    bool know_client = false;
-
-    SOL_DBG("Client Bootstrap Request received");
-
-    response = sol_coap_packet_new(req);
-    SOL_NULL_CHECK(response, -ENOMEM);
-
-    r = extract_bootstrap_client_info(req, &client_name);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-    for (i = 0; server->known_clients[i]; i++) {
-        if (sol_str_slice_str_eq(client_name, server->known_clients[i]))
-            know_client = true;
-    }
-
-    if (!know_client) {
-        SOL_WRN("Client %.*s bootstrap request received, but this Bootstrap Server"
-            " doesn't have Bootstrap Information for this client.",
-            SOL_STR_SLICE_PRINT(client_name));
-        goto err_exit;
-    }
-
-    r = new_bootstrap_client_info(&bs_cinfo, cliaddr, client_name);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-    r = sol_ptr_vector_append(&server->clients, bs_cinfo);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_del_client);
-
-
-    r = sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_CHANGED);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit_del_client_list);
-
-    SOL_DBG("Client %s bootstrap request received."
-        " Bootstrap Process will start now.", bs_cinfo->name);
-
-    r = sol_coap_send_packet(coap, response, cliaddr);
-    dispatch_bootstrap_event_to_server(server, bs_cinfo);
-
-    return r;
-
-err_exit_del_client_list:
-    sol_ptr_vector_remove(&server->clients, bs_cinfo);
-err_exit_del_client:
-    bootstrap_client_info_del(bs_cinfo);
-err_exit:
-    sol_coap_header_set_code(response, SOL_COAP_RESPONSE_CODE_BAD_REQUEST);
-    sol_coap_send_packet(coap, response, cliaddr);
-    return r;
-}
-
-static const struct sol_coap_resource bootstrap_request_interface = {
-    SOL_SET_API_VERSION(.api_version = SOL_COAP_RESOURCE_API_VERSION, )
-    .post = bootstrap_request,
-    .flags = SOL_COAP_FLAGS_NONE,
-    .path = {
-        SOL_STR_SLICE_LITERAL("bs"),
-        SOL_STR_SLICE_EMPTY
-    }
-};
-
-static void
-observer_entry_free(struct observer_entry *entry)
-{
-    sol_monitors_clear(&entry->monitors);
-    free(entry->path);
-    free(entry);
-}
-
-static void
-remove_observer_entry(struct sol_ptr_vector *entries,
-    struct observer_entry *entry)
-{
-    int r;
-
-    r = sol_ptr_vector_del_element(entries, entry);
-    SOL_INT_CHECK(r, < 0);
-    observer_entry_free(entry);
-}
-
-static struct observer_entry *
-find_observer_entry(struct sol_ptr_vector *entries,
-    struct sol_lwm2m_client_info *cinfo, const char *path)
-{
-    uint16_t i;
-    struct observer_entry *entry;
-
-    SOL_PTR_VECTOR_FOREACH_IDX (entries, entry, i) {
-        if (entry->cinfo == cinfo && streq(path, entry->path))
-            return entry;
-    }
-
-    return NULL;
-}
-
-static int
-observer_entry_new(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *cinfo, const char *path,
-    struct observer_entry **entry)
-{
-    int r = -ENOMEM;
-
-    *entry = calloc(1, sizeof(struct observer_entry));
-    SOL_NULL_CHECK(*entry, r);
-
-    (*entry)->path = strdup(path);
-    SOL_NULL_CHECK_GOTO((*entry)->path, err_exit);
-
-    sol_monitors_init(&(*entry)->monitors, NULL);
-    (*entry)->server = server;
-    (*entry)->cinfo = cinfo;
-
-    r = sol_ptr_vector_append(&server->observers, *entry);
-    SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-    return 0;
-err_exit:
-    free((*entry)->path);
-    free(*entry);
-    return r;
-}
-
-static int
-add_to_monitors(struct sol_monitors *monitors, sol_monitors_cb_t cb, const void *data)
-{
-    struct sol_monitors_entry *m;
-
-    SOL_NULL_CHECK(cb, -EINVAL);
-
-    m = sol_monitors_append(monitors, cb, data);
-    SOL_NULL_CHECK(m, -ENOMEM);
-
-    return 0;
-}
-
-static int
-remove_from_monitors(struct sol_monitors *monitors, sol_monitors_cb_t cb, const void *data)
-{
-    int i;
-
-    SOL_NULL_CHECK(cb, -EINVAL);
-
-    i = sol_monitors_find(monitors, cb, data);
-    SOL_INT_CHECK(i, < 0, i);
-
-    return sol_monitors_del(monitors, i);
-}
-
-static int
-observer_entry_add_monitor(struct observer_entry *entry,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    enum sol_coap_response_code response_code,
-    enum sol_lwm2m_content_type content_type,
-    struct sol_str_slice content),
-    const void *data)
-{
-    SOL_NULL_CHECK(entry, -EINVAL);
-
-    return add_to_monitors(&entry->monitors, (sol_monitors_cb_t)cb, data);
-}
-
-static int
-observer_entry_del_monitor(struct observer_entry *entry,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    enum sol_coap_response_code response_code,
-    enum sol_lwm2m_content_type content_type,
-    struct sol_str_slice content),
-    const void *data)
-{
-    SOL_NULL_CHECK(entry, -EINVAL);
-
-    return remove_from_monitors(&entry->monitors, (sol_monitors_cb_t)cb, data);
-}
-
-SOL_API struct sol_lwm2m_server *
-sol_lwm2m_server_new(uint16_t port)
-{
-    struct sol_lwm2m_server *server;
-    struct sol_network_link_addr servaddr = { .family = SOL_NETWORK_FAMILY_INET6,
-                                              .port = port };
-    int r;
-
-    SOL_LOG_INTERNAL_INIT_ONCE;
-
-    server = calloc(1, sizeof(struct sol_lwm2m_server));
-    SOL_NULL_CHECK(server, NULL);
-
-    server->coap = sol_coap_server_new(&servaddr, false);
-    SOL_NULL_CHECK_GOTO(server->coap, err_coap);
-
-    sol_ptr_vector_init(&server->clients);
-    sol_ptr_vector_init(&server->clients_to_delete);
-    sol_ptr_vector_init(&server->observers);
-    sol_monitors_init(&server->registration, NULL);
-
-    r = sol_coap_server_register_resource(server->coap,
-        &registration_interface, server);
-    SOL_INT_CHECK_GOTO(r, < 0, err_register);
-
-    return server;
-
-err_register:
-    sol_coap_server_unref(server->coap);
-err_coap:
-    free(server);
-    return NULL;
-}
-
-SOL_API void
-sol_lwm2m_server_del(struct sol_lwm2m_server *server)
-{
-    uint16_t i;
-    struct sol_lwm2m_client_info *cinfo;
-    struct observer_entry *entry;
-
-    SOL_NULL_CHECK(server);
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->observers, entry, i)
-        entry->removed = true;
-
-    sol_coap_server_unref(server->coap);
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->clients, cinfo, i)
-        client_info_del(cinfo);
-
-    if (server->lifetime_ctx.timeout)
-        sol_timeout_del(server->lifetime_ctx.timeout);
-
-    clients_to_delete_clear(&server->clients_to_delete);
-    sol_monitors_clear(&server->registration);
-    sol_ptr_vector_clear(&server->clients);
-    free(server);
-}
-
-SOL_API int
-sol_lwm2m_server_add_registration_monitor(struct sol_lwm2m_server *server,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *cinfo,
-    enum sol_lwm2m_registration_event event),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-
-    return add_to_monitors(&server->registration, (sol_monitors_cb_t)cb, data);
-}
-
-SOL_API int
-sol_lwm2m_server_del_registration_monitor(struct sol_lwm2m_server *server,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *cinfo,
-    enum sol_lwm2m_registration_event event),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-
-    return remove_from_monitors(&server->registration, (sol_monitors_cb_t)cb, data);
-}
-
-SOL_API struct sol_lwm2m_bootstrap_server *
-sol_lwm2m_bootstrap_server_new(uint16_t port, const char **known_clients)
-{
-    struct sol_lwm2m_bootstrap_server *server;
-    struct sol_network_link_addr servaddr = { .family = SOL_NETWORK_FAMILY_INET6,
-                                              .port = port };
-    int r;
-
-    SOL_LOG_INTERNAL_INIT_ONCE;
-
-    SOL_NULL_CHECK(known_clients, NULL);
-
-    server = calloc(1, sizeof(struct sol_lwm2m_bootstrap_server));
-    SOL_NULL_CHECK(server, NULL);
-
-    server->coap = sol_coap_server_new(&servaddr, false);
-    SOL_NULL_CHECK_GOTO(server->coap, err_coap);
-
-    server->known_clients = known_clients;
-
-    sol_ptr_vector_init(&server->clients);
-
-    sol_monitors_init(&server->bootstrap, NULL);
-
-    r = sol_coap_server_register_resource(server->coap,
-        &bootstrap_request_interface, server);
-    SOL_INT_CHECK_GOTO(r, < 0, err_register);
-
-    return server;
-
-err_register:
-    sol_coap_server_unref(server->coap);
-err_coap:
-    free(server);
-    return NULL;
-}
-
-SOL_API void
-sol_lwm2m_bootstrap_server_del(struct sol_lwm2m_bootstrap_server *server)
-{
-    uint16_t i;
-    struct sol_lwm2m_bootstrap_client_info *bs_cinfo;
-
-    SOL_NULL_CHECK(server);
-
-    sol_coap_server_unref(server->coap);
-
-    SOL_PTR_VECTOR_FOREACH_IDX (&server->clients, bs_cinfo, i)
-        bootstrap_client_info_del(bs_cinfo);
-
-    sol_monitors_clear(&server->bootstrap);
-    sol_ptr_vector_clear(&server->clients);
-    free(server);
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_add_request_monitor(struct sol_lwm2m_bootstrap_server *server,
-    void (*cb)(void *data,
-    struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *bs_cinfo),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-
-    return add_to_monitors(&server->bootstrap, (sol_monitors_cb_t)cb, data);
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_del_request_monitor(struct sol_lwm2m_bootstrap_server *server,
-    void (*cb)(void *data,
-    struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *bs_cinfo),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-
-    return remove_from_monitors(&server->bootstrap, (sol_monitors_cb_t)cb, data);
 }
 
 SOL_API int
@@ -1493,1274 +168,6 @@ sol_lwm2m_client_del_bootstrap_finish_monitor(struct sol_lwm2m_client *client,
     SOL_NULL_CHECK(client, -EINVAL);
 
     return remove_from_monitors(&client->bootstrap, (sol_monitors_cb_t)cb, data);
-}
-
-SOL_API const struct sol_ptr_vector *
-sol_lwm2m_server_get_clients(const struct sol_lwm2m_server *server)
-{
-    SOL_NULL_CHECK(server, NULL);
-
-    return &server->clients;
-}
-
-SOL_API const char *
-sol_lwm2m_bootstrap_client_info_get_name(const struct sol_lwm2m_bootstrap_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return client->name;
-}
-
-SOL_API const char *
-sol_lwm2m_client_info_get_name(const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return client->name;
-}
-
-SOL_API const char *
-sol_lwm2m_client_info_get_location(const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return client->location;
-}
-
-SOL_API const char *
-sol_lwm2m_client_info_get_sms_number(const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return client->sms;
-}
-
-SOL_API const char *
-sol_lwm2m_client_info_get_objects_path(
-    const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return client->objects_path;
-}
-
-SOL_API int
-sol_lwm2m_client_info_get_lifetime(const struct sol_lwm2m_client_info *client,
-    uint32_t *lifetime)
-{
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(lifetime, -EINVAL);
-
-    *lifetime = client->lifetime;
-    return 0;
-}
-
-SOL_API enum sol_lwm2m_binding_mode
-sol_lwm2m_client_info_get_binding_mode(
-    const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, SOL_LWM2M_BINDING_MODE_UNKNOWN);
-
-    return client->binding;
-}
-
-SOL_API const struct sol_network_link_addr *
-sol_lwm2m_client_info_get_address(const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return &client->cliaddr;
-}
-
-SOL_API const struct sol_network_link_addr *
-sol_lwm2m_bootstrap_client_info_get_address(const struct sol_lwm2m_bootstrap_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return &client->cliaddr;
-}
-
-SOL_API const struct sol_ptr_vector *
-sol_lwm2m_client_info_get_objects(const struct sol_lwm2m_client_info *client)
-{
-    SOL_NULL_CHECK(client, NULL);
-
-    return &client->objects;
-}
-
-SOL_API int
-sol_lwm2m_client_object_get_id(const struct sol_lwm2m_client_object *object,
-    uint16_t *id)
-{
-    SOL_NULL_CHECK(object, -EINVAL);
-    SOL_NULL_CHECK(id, -EINVAL);
-
-    *id = object->id;
-    return 0;
-}
-
-SOL_API const struct sol_ptr_vector *
-sol_lwm2m_client_object_get_instances(
-    const struct sol_lwm2m_client_object *object)
-{
-    SOL_NULL_CHECK(object, NULL);
-
-    return &object->instances;
-}
-
-static size_t
-get_int_size(int64_t i)
-{
-    if (i >= INT8_MIN && i <= INT8_MAX)
-        return 1;
-    if (i >= INT16_MIN && i <= INT16_MAX)
-        return 2;
-    if (i >= INT32_MIN && i <= INT32_MAX)
-        return 4;
-    return 8;
-}
-
-static int
-get_resource_len(const struct sol_lwm2m_resource *resource, uint16_t index,
-    size_t *len)
-{
-    switch (resource->data_type) {
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_STRING:
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE:
-        *len = resource->data[index].content.blob->size;
-        return 0;
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_INT:
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_TIME:
-        *len = get_int_size(resource->data[index].content.integer);
-        return 0;
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL:
-        *len = 1;
-        return 0;
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT:
-        *len = 8;
-        return 0;
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK:
-        *len = OBJ_LINK_LEN;
-        return 0;
-    default:
-        return -EINVAL;
-    }
-}
-
-static void
-swap_bytes(uint8_t *to_swap, size_t len)
-{
-#if __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
-    return;
-#elif __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-    uint8_t swap;
-    size_t i, j, stop;
-
-    stop = len / 2;
-
-    for (i = 0, j = len - 1; i != stop; i++, j--) {
-        swap = to_swap[i];
-        to_swap[i] = to_swap[j];
-        to_swap[j] = swap;
-    }
-#else
-#error "Unknown byte order"
-#endif
-}
-
-static int
-add_float_resource(struct sol_buffer *buf, double fp, size_t len)
-{
-    uint8_t *bytes = NULL;
-    float f;
-    double d;
-
-    if (len == 4) {
-        f = (float)fp;
-        swap_bytes((uint8_t *)&f, len);
-        bytes = (uint8_t *)&f;
-    } else {
-        d = fp;
-        swap_bytes((uint8_t *)&d, len);
-        bytes = (uint8_t *)&d;
-    }
-
-    return sol_buffer_append_bytes(buf, bytes, len);
-}
-
-static int
-add_int_resource(struct sol_buffer *buf, int64_t i, size_t len)
-{
-    swap_bytes((uint8_t *)&i, len);
-    return sol_buffer_append_bytes(buf, (uint8_t *)&i, len);
-}
-
-static int
-add_resource_bytes_to_buffer(const struct sol_lwm2m_resource *resource,
-    struct sol_buffer *buf, uint16_t idx)
-{
-    int r;
-    uint8_t b;
-    size_t len;
-
-    r = get_resource_len(resource, idx, &len);
-    SOL_INT_CHECK(r, < 0, r);
-
-    switch (resource->data_type) {
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_STRING:
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE:
-        return sol_buffer_append_slice(buf, sol_str_slice_from_blob(resource->data[idx].content.blob));
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_INT:
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_TIME:
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK:
-        return add_int_resource(buf, resource->data[idx].content.integer, len);
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL:
-        b = resource->data[idx].content.integer != 0 ? 1 : 0;
-        return sol_buffer_append_bytes(buf, (uint8_t *)&b, 1);
-    case SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT:
-        return add_float_resource(buf, resource->data[idx].content.fp, len);
-    default:
-        return -EINVAL;
-    }
-}
-
-static int
-set_packet_payload(struct sol_coap_packet *pkt,
-    const uint8_t *data, uint16_t len)
-{
-    struct sol_buffer *buf;
-    int r;
-
-    r = sol_coap_packet_get_payload(pkt, &buf, NULL);
-    SOL_INT_CHECK(r, < 0, r);
-
-    return sol_buffer_append_bytes(buf, data, len);
-}
-
-static int
-setup_tlv_header(enum sol_lwm2m_tlv_type tlv_type, uint16_t res_id,
-    struct sol_buffer *buf, size_t data_len)
-{
-    int r;
-    uint8_t tlv_data[6];
-    size_t tlv_data_len;
-
-    tlv_data_len = 2;
-
-    tlv_data[0] = tlv_type;
-
-    if (res_id > UINT8_MAX) {
-        tlv_data[0] |= ID_HAS_16BITS_MASK;
-        tlv_data[1] = (res_id >> 8) & 255;
-        tlv_data[2] = res_id & 255;
-        tlv_data_len++;
-    } else
-        tlv_data[1] = res_id;
-
-    if (data_len <= 7)
-        tlv_data[0] |= data_len;
-    else if (data_len <= UINT8_MAX) {
-        tlv_data[tlv_data_len++] = data_len;
-        tlv_data[0] |= LEN_IS_8BITS_MASK;
-    } else if (data_len <= UINT16_MAX) {
-        tlv_data[tlv_data_len++] = (data_len >> 8) & 255;
-        tlv_data[tlv_data_len++] = data_len & 255;
-        tlv_data[0] |= LEN_IS_16BITS_MASK;
-    } else if (data_len <= UINT24_MAX) {
-        tlv_data[tlv_data_len++] = (data_len >> 16) & 255;
-        tlv_data[tlv_data_len++] = (data_len >> 8) & 255;
-        tlv_data[tlv_data_len++] = data_len & 255;
-        tlv_data[0] |= LEN_IS_24BITS_MASK;
-    } else
-        return -ENOMEM;
-
-    r = sol_buffer_append_bytes(buf, tlv_data, tlv_data_len);
-    SOL_INT_CHECK(r, < 0, r);
-
-    return 0;
-}
-
-static int
-setup_tlv(struct sol_lwm2m_resource *resource, struct sol_buffer *buf)
-{
-    int r;
-    size_t data_len, len;
-    uint16_t i;
-    enum sol_lwm2m_tlv_type type;
-
-    LWM2M_RESOURCE_CHECK_API(resource, -EINVAL);
-
-    for (i = 0, data_len = 0; i < resource->data_len; i++) {
-        r = get_resource_len(resource, i, &len);
-        SOL_INT_CHECK(r, < 0, r);
-        data_len += len;
-    }
-
-    switch (resource->type) {
-    case SOL_LWM2M_RESOURCE_TYPE_SINGLE:
-        type = SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE;
-        break;
-    case SOL_LWM2M_RESOURCE_TYPE_MULTIPLE:
-        type = SOL_LWM2M_TLV_TYPE_MULTIPLE_RESOURCES;
-        data_len += (resource->data_len * 2);
-        break;
-    default:
-        SOL_WRN("Unknown resource type '%d'", (int)resource->type);
-        return -EINVAL;
-    }
-
-    r = setup_tlv_header(type, resource->id, buf, data_len);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (type == SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE)
-        return add_resource_bytes_to_buffer(resource, buf, 0);
-
-    for (i = 0; i < resource->data_len; i++) {
-        r = get_resource_len(resource, i, &data_len);
-        SOL_INT_CHECK(r, < 0, r);
-        r = setup_tlv_header(SOL_LWM2M_TLV_TYPE_RESOURCE_INSTANCE, resource->data[i].id,
-            buf, data_len);
-        SOL_INT_CHECK(r, < 0, r);
-        r = add_resource_bytes_to_buffer(resource, buf, i);
-        SOL_INT_CHECK(r, < 0, r);
-    }
-
-    return 0;
-}
-
-static int
-resources_to_tlv(struct sol_lwm2m_resource *resources,
-    size_t len, struct sol_buffer *tlvs)
-{
-    int r;
-    size_t i;
-
-    for (i = 0; i < len; i++) {
-        r = setup_tlv(&resources[i], tlvs);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    return 0;
-
-exit:
-    return r;
-}
-
-static int
-instances_to_tlv(struct sol_lwm2m_resource **instances,
-    size_t *instances_len, uint16_t *instances_ids, size_t len, struct sol_buffer *tlvs)
-{
-    int r;
-    size_t i, j, res, instance_data_len, data_len;
-
-    for (i = 0; i < len; i++) {
-        for (res = 0, instance_data_len = 0; res < instances_len[i]; res++)
-            for (j = 0; j < instances[i][res].data_len; j++) {
-                r = get_resource_len(&instances[i][res], j, &data_len);
-                SOL_INT_CHECK(r, < 0, r);
-
-                instance_data_len += data_len;
-            }
-
-        r = setup_tlv_header(SOL_LWM2M_TLV_TYPE_OBJECT_INSTANCE,
-            instances_ids[i], tlvs, instance_data_len);
-        SOL_INT_CHECK(r, < 0, r);
-
-        r = resources_to_tlv(instances[i], instances_len[i], tlvs);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    return 0;
-
-exit:
-    return r;
-}
-
-static int
-add_coap_int_option(struct sol_coap_packet *pkt,
-    enum sol_coap_option opt, const void *data, uint16_t len)
-{
-    uint8_t buf[sizeof(int64_t)] = { };
-
-    memcpy(buf, data, len);
-    swap_bytes(buf, len);
-    return sol_coap_add_option(pkt, opt, buf, len);
-}
-
-static int
-get_coap_int_option(struct sol_coap_packet *pkt,
-    enum sol_coap_option opt, uint16_t *value)
-{
-    const void *v;
-    uint16_t len;
-
-    v = sol_coap_find_first_option(pkt, opt, &len);
-
-    if (!v)
-        return -ENOENT;
-
-    memcpy(value, v, len);
-    swap_bytes((uint8_t *)value, len);
-    return 0;
-}
-
-static int
-setup_coap_packet(enum sol_coap_method method,
-    enum sol_coap_message_type type, const char *objects_path, const char *path,
-    uint8_t *obs, int64_t *token, struct sol_lwm2m_resource *resources,
-    struct sol_lwm2m_resource **instances, size_t *instances_len,
-    uint16_t *instances_ids, size_t len,
-    const char *execute_args,
-    struct sol_coap_packet **pkt)
-{
-    struct sol_buffer buf = SOL_BUFFER_INIT_EMPTY;
-    struct sol_buffer tlvs =
-        SOL_BUFFER_INIT_FLAGS(NULL, 0, SOL_BUFFER_FLAGS_NO_NUL_BYTE);
-    struct sol_random *random;
-    uint16_t content_type, content_len = 0;
-    const uint8_t *content_data;
-    int64_t t;
-    int r;
-
-    random = sol_random_new(SOL_RANDOM_DEFAULT, 0);
-    SOL_NULL_CHECK(random, -ENOMEM);
-
-    *pkt = sol_coap_packet_new_request(method, type);
-    r = -ENOMEM;
-    SOL_NULL_CHECK_GOTO(*pkt, exit);
-
-    r = sol_random_get_int64(random, &t);
-    if (r < 0) {
-        SOL_WRN("Could not generate a random number");
-        goto exit;
-    }
-
-    r = sol_coap_header_set_token(*pkt, (uint8_t *)&t,
-        (uint8_t)sizeof(int64_t));
-    if (r < 0) {
-        SOL_WRN("Could not set the token");
-        goto exit;
-    }
-    SOL_DBG("Setting token as %" PRId64 ", len = %zu", t, sizeof(int64_t));
-
-    if (token)
-        *token = t;
-
-    if (obs) {
-        r = add_coap_int_option(*pkt, SOL_COAP_OPTION_OBSERVE, obs,
-            sizeof(uint8_t));
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    if (objects_path) {
-        r = sol_buffer_append_slice(&buf,
-            sol_str_slice_from_str(objects_path));
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    r = sol_buffer_append_slice(&buf, sol_str_slice_from_str(path));
-    SOL_INT_CHECK_GOTO(r, < 0, exit);
-
-    if (strcmp(path, "/") != 0) {
-        r = sol_coap_packet_add_uri_path_option(*pkt, buf.data);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    if (execute_args) {
-        size_t str_len;
-        content_type = SOL_LWM2M_CONTENT_TYPE_TEXT;
-        content_data = (const uint8_t *)execute_args;
-        str_len = strlen(execute_args);
-        r = -ENOMEM;
-        SOL_INT_CHECK_GOTO(str_len, >= UINT16_MAX, exit);
-        content_len = str_len;
-    } else if (resources) {
-        content_type = SOL_LWM2M_CONTENT_TYPE_TLV;
-        r = resources_to_tlv(resources, len, &tlvs);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-        r = -ENOMEM;
-        SOL_INT_CHECK_GOTO(tlvs.used, >= UINT16_MAX, exit);
-        content_data = tlvs.data;
-        content_len = tlvs.used;
-    } else if (instances) {
-        content_type = SOL_LWM2M_CONTENT_TYPE_TLV;
-        r = instances_to_tlv(instances, instances_len, instances_ids, len, &tlvs);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-        r = -ENOMEM;
-        SOL_INT_CHECK_GOTO(tlvs.used, >= UINT16_MAX, exit);
-        content_data = tlvs.data;
-        content_len = tlvs.used;
-    }
-
-    if (content_len > 0) {
-        r = add_coap_int_option(*pkt, SOL_COAP_OPTION_CONTENT_FORMAT,
-            &content_type, sizeof(content_type));
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-
-        r = set_packet_payload(*pkt, content_data, content_len);
-        SOL_INT_CHECK_GOTO(r, < 0, exit);
-    }
-
-    r = 0;
-
-exit:
-    if (r < 0)
-        sol_coap_packet_unref(*pkt);
-    sol_buffer_fini(&tlvs);
-    sol_buffer_fini(&buf);
-    sol_random_del(random);
-    return r;
-}
-
-static void
-extract_content(struct sol_coap_packet *req, uint8_t *code,
-    enum sol_lwm2m_content_type *type, struct sol_str_slice *content)
-{
-    struct sol_buffer *buf;
-    size_t offset;
-    int r;
-
-    r = sol_coap_header_get_code(req, code);
-    SOL_INT_CHECK(r, < 0);
-
-    if (!sol_coap_packet_has_payload(req))
-        return;
-
-    r = sol_coap_packet_get_payload(req, &buf, &offset);
-    SOL_INT_CHECK(r, < 0);
-    content->len = buf->used - offset;
-    content->data = sol_buffer_at(buf, offset);
-    r = get_coap_int_option(req, SOL_COAP_OPTION_CONTENT_FORMAT,
-        (uint16_t *)type);
-    if (r < 0)
-        SOL_INF("Content format not specified");
-}
-
-static bool
-observation_request_reply(void *data, struct sol_coap_server *coap_server,
-    struct sol_coap_packet *req, const struct sol_network_link_addr *cliaddr)
-{
-    struct observer_entry *entry = data;
-    struct sol_monitors_entry *m;
-    struct sol_str_slice content = SOL_STR_SLICE_EMPTY;
-    enum sol_lwm2m_content_type type = SOL_LWM2M_CONTENT_TYPE_TEXT;
-    uint16_t i;
-    uint8_t code = SOL_COAP_RESPONSE_CODE_GATEWAY_TIMEOUT;
-    bool keep_alive = true;
-
-    if (!cliaddr && !req) {
-        //Cancel observation
-        if (entry->removed) {
-            remove_observer_entry(&entry->server->observers, entry);
-            return false;
-        }
-        SOL_WRN("Could not complete the observation request on client:%s"
-            " path:%s", entry->cinfo->name, entry->path);
-        keep_alive = false;
-    } else {
-        extract_content(req, &code, &type, &content);
-        send_ack_if_needed(coap_server, req, cliaddr);
-    }
-
-    SOL_MONITORS_WALK (&entry->monitors, m, i)
-        ((void (*)(void *, struct sol_lwm2m_server *,
-        struct sol_lwm2m_client_info *,
-        const char *,
-        enum sol_coap_response_code,
-        enum sol_lwm2m_content_type,
-        struct sol_str_slice))m->cb)((void *)m->data, entry->server,
-            entry->cinfo, entry->path, code, type, content);
-
-    return keep_alive;
-}
-
-SOL_API int
-sol_lwm2m_server_add_observer(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    enum sol_coap_response_code response_code,
-    enum sol_lwm2m_content_type content_type,
-    struct sol_str_slice content),
-    const void *data)
-{
-    struct observer_entry *entry;
-    struct sol_coap_packet *pkt;
-    uint8_t obs = 0;
-    int r;
-    bool send_msg = false;
-
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-
-    entry = find_observer_entry(&server->observers, client, path);
-
-    if (!entry) {
-        send_msg = true;
-        r = observer_entry_new(server, client, path, &entry);
-        SOL_INT_CHECK(r, < 0, r);
-    }
-
-    r = observer_entry_add_monitor(entry, cb, data);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (!send_msg)
-        return 0;
-
-    r = setup_coap_packet(SOL_COAP_METHOD_GET, SOL_COAP_MESSAGE_TYPE_CON,
-        client->objects_path, path, &obs, &entry->token, NULL, NULL, NULL, NULL, 0, NULL, &pkt);
-    SOL_INT_CHECK(r, < 0, r);
-
-    return sol_coap_send_packet_with_reply(server->coap, pkt, &client->cliaddr,
-        observation_request_reply, entry);
-}
-
-SOL_API int
-sol_lwm2m_server_del_observer(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    enum sol_coap_response_code response_code,
-    enum sol_lwm2m_content_type content_type,
-    struct sol_str_slice content),
-    const void *data)
-{
-    struct observer_entry *entry;
-    int r;
-    int64_t token;
-
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-
-    entry = find_observer_entry(&server->observers, client, path);
-    SOL_NULL_CHECK(entry, -ENOENT);
-
-    r = observer_entry_del_monitor(entry, cb, data);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (entry->monitors.entries.len)
-        return 0;
-
-    entry->removed = true;
-    token = entry->token;
-
-    return sol_coap_unobserve_by_token(server->coap, &entry->cinfo->cliaddr,
-        (uint8_t *)&token, sizeof(token));
-}
-
-SOL_API int
-sol_lwm2m_resource_init(struct sol_lwm2m_resource *resource,
-    uint16_t id, enum sol_lwm2m_resource_type type, uint16_t resource_len,
-    enum sol_lwm2m_resource_data_type data_type, ...)
-{
-    uint16_t i;
-    va_list ap;
-    struct sol_blob *blob;
-    int r = -EINVAL;
-
-    SOL_NULL_CHECK(resource, -EINVAL);
-    SOL_INT_CHECK(data_type, == SOL_LWM2M_RESOURCE_DATA_TYPE_NONE, -EINVAL);
-    SOL_INT_CHECK(resource_len, <= 0, -EINVAL);
-
-    LWM2M_RESOURCE_CHECK_API(resource, -EINVAL);
-
-    resource->id = id;
-    resource->type = type;
-    resource->data_type = data_type;
-    resource->data = calloc(resource_len, sizeof(struct sol_lwm2m_resource_data));
-    SOL_NULL_CHECK(resource->data, -ENOMEM);
-    resource->data_len = resource_len;
-
-    va_start(ap, data_type);
-
-    for (i = 0; i < resource_len; i++) {
-        if (resource->type == SOL_LWM2M_RESOURCE_TYPE_MULTIPLE)
-            resource->data[i].id = va_arg(ap, int);
-
-        switch (resource->data_type) {
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE:
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_STRING:
-            blob = va_arg(ap, struct sol_blob *);
-            SOL_NULL_CHECK_GOTO(blob, err_exit);
-            resource->data[i].content.blob = sol_blob_ref(blob);
-            SOL_NULL_CHECK_GOTO(resource->data[i].content.blob, err_ref);
-            break;
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT:
-            resource->data[i].content.fp = va_arg(ap, double);
-            break;
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_INT:
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_TIME:
-            resource->data[i].content.integer = va_arg(ap, int64_t);
-            break;
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL:
-            resource->data[i].content.integer = va_arg(ap, int);
-            break;
-        case SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK:
-            resource->data[i].content.integer = (uint16_t)va_arg(ap, int);
-            resource->data[i].content.integer = (resource->data[i].content.integer << 16) |
-                (uint16_t)va_arg(ap, int);
-            break;
-        default:
-            SOL_WRN("Unknown resource data type");
-            goto err_exit;
-        }
-    }
-
-    va_end(ap);
-    return 0;
-
-err_ref:
-    r = -EOVERFLOW;
-err_exit:
-    if (data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE ||
-        data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_STRING) {
-        uint16_t until = i;
-
-        for (i = 0; i < until; i++)
-            sol_blob_unref(resource->data[i].content.blob);
-    }
-    free(resource->data);
-    va_end(ap);
-    return r;
-}
-
-SOL_API int
-sol_lwm2m_resource_init_vector(struct sol_lwm2m_resource *resource,
-    uint16_t id, enum sol_lwm2m_resource_data_type data_type,
-    struct sol_vector *res_instances)
-{
-    uint16_t i;
-    int r = -EINVAL;
-
-    SOL_NULL_CHECK(resource, -EINVAL);
-    SOL_INT_CHECK(data_type, == SOL_LWM2M_RESOURCE_DATA_TYPE_NONE, -EINVAL);
-    SOL_NULL_CHECK(res_instances, -EINVAL);
-    SOL_INT_CHECK(res_instances->len, <= 0, -EINVAL);
-
-    LWM2M_RESOURCE_CHECK_API(resource, -EINVAL);
-
-    resource->id = id;
-    resource->type = SOL_LWM2M_RESOURCE_TYPE_MULTIPLE;
-    resource->data_type = data_type;
-    resource->data = calloc(res_instances->len, sizeof(struct sol_lwm2m_resource_data));
-    SOL_NULL_CHECK(resource->data, -ENOMEM);
-    resource->data_len = res_instances->len;
-
-    for (i = 0; i < res_instances->len; i++) {
-        void *v = sol_vector_get_no_check(res_instances, i);
-        struct sol_lwm2m_resource_data *res_data = v;
-        resource->data[i].id = res_data->id;
-
-        if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE ||
-            resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_STRING) {
-            struct sol_blob *blob = res_data->content.blob;
-            SOL_NULL_CHECK_GOTO(blob, err_exit);
-            resource->data[i].content.blob = sol_blob_ref(blob);
-            SOL_NULL_CHECK_GOTO(resource->data[i].content.blob, err_ref);
-
-        } else if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT) {
-            resource->data[i].content.fp = res_data->content.fp;
-
-        } else if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_INT ||
-            resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_TIME) {
-            resource->data[i].content.integer = res_data->content.integer;
-
-        } else if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL) {
-            resource->data[i].content.integer = res_data->content.integer;
-
-        } else if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK) {
-            resource->data[i].content.integer = (uint16_t)res_data->content.integer;
-            resource->data[i].content.integer = (resource->data[i].content.integer << 16) |
-                (uint16_t)res_data->content.integer;
-
-        } else {
-            SOL_WRN("Unknown resource data type");
-            goto err_exit;
-        }
-    }
-
-    return 0;
-
-err_ref:
-    r = -EOVERFLOW;
-err_exit:
-    if (data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE ||
-        data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_STRING) {
-        uint16_t until = i;
-
-        for (i = 0; i < until; i++)
-            sol_blob_unref(resource->data[i].content.blob);
-    }
-    free(resource->data);
-    return r;
-}
-
-static bool
-management_reply(void *data, struct sol_coap_server *server,
-    struct sol_coap_packet *req, const struct sol_network_link_addr *cliaddr)
-{
-    struct management_ctx *ctx = data;
-    uint8_t code = 0;
-    enum sol_lwm2m_content_type content_type = SOL_LWM2M_CONTENT_TYPE_TEXT;
-    struct sol_str_slice content = SOL_STR_SLICE_EMPTY;
-
-    if (!cliaddr && !req)
-        code = SOL_COAP_RESPONSE_CODE_GATEWAY_TIMEOUT;
-
-    switch (ctx->type) {
-    case MANAGEMENT_DELETE:
-    case MANAGEMENT_CREATE:
-    case MANAGEMENT_WRITE:
-    case MANAGEMENT_EXECUTE:
-        if (!code)
-            sol_coap_header_get_code(req, &code);
-        ((void (*)(void *,
-        struct sol_lwm2m_server *,
-        struct sol_lwm2m_client_info *, const char *,
-        enum sol_coap_response_code))ctx->cb)
-            ((void *)ctx->data, ctx->server, ctx->cinfo, ctx->path, code);
-        break;
-    default: //Read
-        if (!code)
-            extract_content(req, &code, &content_type, &content);
-        ((void (*)(void *, struct sol_lwm2m_server *,
-        struct sol_lwm2m_client_info *,
-        const char *,
-        enum sol_coap_response_code,
-        enum sol_lwm2m_content_type,
-        struct sol_str_slice))ctx->cb)((void *)ctx->data, ctx->server,
-            ctx->cinfo, ctx->path, code, content_type, content);
-        break;
-    }
-
-    if (code != SOL_COAP_RESPONSE_CODE_GATEWAY_TIMEOUT)
-        send_ack_if_needed(server, req, cliaddr);
-    free(ctx->path);
-    free(ctx);
-    return false;
-}
-
-static int
-send_management_packet(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    enum management_type type, void *cb, const void *data,
-    enum sol_coap_method method,
-    struct sol_lwm2m_resource *resources, size_t len, const char *execute_args)
-{
-    int r;
-    struct sol_coap_packet *pkt;
-    struct management_ctx *ctx;
-
-    r = setup_coap_packet(method, SOL_COAP_MESSAGE_TYPE_CON,
-        client->objects_path, path, NULL, NULL, resources, NULL, NULL, NULL,
-        len, execute_args, &pkt);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (!cb)
-        return sol_coap_send_packet(server->coap, pkt, &client->cliaddr);
-
-    ctx = malloc(sizeof(struct management_ctx));
-    SOL_NULL_CHECK_GOTO(ctx, err_exit);
-
-    ctx->path = strdup(path);
-    SOL_NULL_CHECK_GOTO(ctx->path, err_exit);
-    ctx->type = type;
-    ctx->server = server;
-    ctx->cinfo = client;
-    ctx->data = data;
-    ctx->cb = cb;
-
-    return sol_coap_send_packet_with_reply(server->coap, pkt, &client->cliaddr,
-        management_reply, ctx);
-
-err_exit:
-    free(ctx);
-    sol_coap_packet_unref(pkt);
-    return -ENOMEM;
-}
-
-/**
- * This checks if the path has the following form: /2/0/1
- */
-static bool
-is_resource_set(const char *path)
-{
-    size_t i;
-    uint8_t slashes;
-    const char *last_slash;
-
-    for (i = 0, slashes = 0; path[i]; i++) {
-        if (path[i] == '/') {
-            last_slash = path + i;
-            slashes++;
-        }
-    }
-
-    if (slashes < 3 || *(last_slash + 1) == '\0')
-        return false;
-    return true;
-}
-
-SOL_API int
-sol_lwm2m_server_write(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    struct sol_lwm2m_resource *resources, size_t len,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    enum sol_coap_method method = SOL_COAP_METHOD_PUT;
-
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-    SOL_NULL_CHECK(resources, -EINVAL);
-
-    if (!is_resource_set(path))
-        method = SOL_COAP_METHOD_POST;
-
-    return send_management_packet(server, client, path,
-        MANAGEMENT_WRITE, cb, data, method, resources, len, NULL);
-}
-
-SOL_API int
-sol_lwm2m_server_execute_resource(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path, const char *args,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_management_packet(server, client, path,
-        MANAGEMENT_EXECUTE, cb, data, SOL_COAP_METHOD_POST, NULL, 0, args);
-}
-
-SOL_API int
-sol_lwm2m_server_delete_object_instance(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_management_packet(server, client, path,
-        MANAGEMENT_DELETE, cb, data, SOL_COAP_METHOD_DELETE, NULL, 0, NULL);
-}
-
-SOL_API int
-sol_lwm2m_server_create_object_instance(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    struct sol_lwm2m_resource *resources, size_t len,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_management_packet(server, client, path,
-        MANAGEMENT_CREATE, cb, data, SOL_COAP_METHOD_POST, resources,
-        len, NULL);
-}
-
-SOL_API int
-sol_lwm2m_server_read(struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    void (*cb)(void *data,
-    struct sol_lwm2m_server *server,
-    struct sol_lwm2m_client_info *client,
-    const char *path,
-    enum sol_coap_response_code response_code,
-    enum sol_lwm2m_content_type content_type,
-    struct sol_str_slice content),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-    SOL_NULL_CHECK(cb, -EINVAL);
-
-    return send_management_packet(server, client, path,
-        MANAGEMENT_READ, cb, data, SOL_COAP_METHOD_GET, NULL, 0, NULL);
-}
-
-static void
-tlv_clear(struct sol_lwm2m_tlv *tlv)
-{
-    LWM2M_TLV_CHECK_API(tlv);
-    sol_buffer_fini(&tlv->content);
-}
-
-SOL_API void
-sol_lwm2m_tlv_clear(struct sol_lwm2m_tlv *tlv)
-{
-    SOL_NULL_CHECK(tlv);
-    tlv_clear(tlv);
-}
-
-SOL_API void
-sol_lwm2m_tlv_list_clear(struct sol_vector *tlvs)
-{
-    uint16_t i;
-    struct sol_lwm2m_tlv *tlv;
-
-    SOL_NULL_CHECK(tlvs);
-
-    SOL_VECTOR_FOREACH_IDX (tlvs, tlv, i)
-        tlv_clear(tlv);
-    sol_vector_clear(tlvs);
-}
-
-SOL_API int
-sol_lwm2m_parse_tlv(const struct sol_str_slice content, struct sol_vector *out)
-{
-    size_t i, offset;
-    struct sol_lwm2m_tlv *tlv;
-    int r;
-
-    SOL_NULL_CHECK(out, -EINVAL);
-
-    sol_vector_init(out, sizeof(struct sol_lwm2m_tlv));
-
-    for (i = 0; i < content.len;) {
-        struct sol_str_slice tlv_content;
-        tlv = sol_vector_append(out);
-        r = -ENOMEM;
-        SOL_NULL_CHECK_GOTO(tlv, err_exit);
-
-        sol_buffer_init(&tlv->content);
-
-        SOL_SET_API_VERSION(tlv->api_version = SOL_LWM2M_TLV_API_VERSION; )
-
-        tlv->type = content.data[i] & TLV_TYPE_MASK;
-
-        if ((content.data[i] & TLV_ID_SIZE_MASK) != TLV_ID_SIZE_MASK) {
-            tlv->id = content.data[i + 1];
-            offset = i + 2;
-        } else {
-            tlv->id = (content.data[i + 1] << 8) | content.data[i + 2];
-            offset = i + 3;
-        }
-
-        SOL_INT_CHECK_GOTO(offset, >= content.len, err_would_overflow);
-
-        switch (content.data[i] & TLV_CONTENT_LENGTH_MASK) {
-        case LENGTH_SIZE_24_BITS:
-            tlv_content.len = (content.data[offset] << 16) |
-                (content.data[offset + 1] << 8) | content.data[offset + 2];
-            offset += 3;
-            break;
-        case LENGTH_SIZE_16_BITS:
-            tlv_content.len = (content.data[offset] << 8) |
-                content.data[offset + 1];
-            offset += 2;
-            break;
-        case LENGTH_SIZE_8_BITS:
-            tlv_content.len = content.data[offset];
-            offset++;
-            break;
-        default:
-            tlv_content.len = content.data[i] & TLV_CONTENT_LENGHT_CUSTOM_MASK;
-        }
-
-        SOL_INT_CHECK_GOTO(offset, >= content.len, err_would_overflow);
-
-        tlv_content.data = content.data + offset;
-
-        r = sol_buffer_append_slice(&tlv->content, tlv_content);
-        SOL_INT_CHECK_GOTO(r, < 0, err_exit);
-
-        SOL_DBG("tlv type: %u, ID: %" PRIu16 ", Size: %zu, Content: %.*s",
-            tlv->type, tlv->id, tlv_content.len,
-            SOL_STR_SLICE_PRINT(tlv_content));
-
-        if (tlv->type != SOL_LWM2M_TLV_TYPE_MULTIPLE_RESOURCES &&
-            tlv->type != SOL_LWM2M_TLV_TYPE_OBJECT_INSTANCE)
-            i += ((offset - i) + tlv_content.len);
-        else
-            i += (offset - i);
-    }
-
-    return 0;
-
-err_would_overflow:
-    r = -EOVERFLOW;
-err_exit:
-    sol_lwm2m_tlv_list_clear(out);
-    return r;
-}
-
-static int
-is_resource(struct sol_lwm2m_tlv *tlv)
-{
-    if (tlv->type != SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE &&
-        tlv->type != SOL_LWM2M_TLV_TYPE_RESOURCE_INSTANCE)
-        return -EINVAL;
-    return 0;
-}
-
-SOL_API int
-sol_lwm2m_tlv_get_int(struct sol_lwm2m_tlv *tlv, int64_t *value)
-{
-    int8_t i8;
-    int16_t i16;
-    int32_t i32;
-    int64_t i64;
-
-    SOL_NULL_CHECK(tlv, -EINVAL);
-    SOL_NULL_CHECK(value, -EINVAL);
-    SOL_INT_CHECK(is_resource(tlv), < 0, -EINVAL);
-    LWM2M_TLV_CHECK_API(tlv, -EINVAL);
-
-#define TO_LOCAL_INT_VALUE(_network_int, _local_int, _out) \
-    memcpy(&(_local_int), _network_int, sizeof((_local_int))); \
-    swap_bytes((uint8_t *)&(_local_int), sizeof((_local_int))); \
-    *(_out) = _local_int;
-
-    switch (tlv->content.used) {
-    case 1:
-        TO_LOCAL_INT_VALUE(tlv->content.data, i8, value);
-        break;
-    case 2:
-        TO_LOCAL_INT_VALUE(tlv->content.data, i16, value);
-        break;
-    case 4:
-        TO_LOCAL_INT_VALUE(tlv->content.data, i32, value);
-        break;
-    case 8:
-        TO_LOCAL_INT_VALUE(tlv->content.data, i64, value);
-        break;
-    default:
-        SOL_WRN("Invalid int size: %zu", tlv->content.used);
-        return -EINVAL;
-    }
-
-    SOL_DBG("TLV has integer data. Value: %" PRId64 "", *value);
-    return 0;
-
-#undef TO_LOCAL_INT_VALUE
-}
-
-SOL_API int
-sol_lwm2m_tlv_get_bool(struct sol_lwm2m_tlv *tlv, bool *value)
-{
-    char v;
-
-    SOL_NULL_CHECK(tlv, -EINVAL);
-    SOL_NULL_CHECK(value, -EINVAL);
-    SOL_INT_CHECK(is_resource(tlv), < 0, -EINVAL);
-    LWM2M_TLV_CHECK_API(tlv, -EINVAL);
-    SOL_INT_CHECK(tlv->content.used, != 1, -EINVAL);
-
-    v = (((char *)tlv->content.data))[0];
-
-    if (v != 0 && v != 1) {
-        SOL_WRN("The TLV value is not '0' or '1'. Actual value:%d", v);
-        return -EINVAL;
-    }
-
-    *value = (bool)v;
-    SOL_DBG("TLV data as bool: %d", (int)*value);
-    return 0;
-}
-
-SOL_API int
-sol_lwm2m_tlv_get_float(struct sol_lwm2m_tlv *tlv, double *value)
-{
-    SOL_NULL_CHECK(tlv, -EINVAL);
-    SOL_NULL_CHECK(value, -EINVAL);
-    SOL_INT_CHECK(is_resource(tlv), < 0, -EINVAL);
-    LWM2M_TLV_CHECK_API(tlv, -EINVAL);
-
-    if (tlv->content.used == 4) {
-        float f;
-        memcpy(&f, tlv->content.data, sizeof(float));
-        swap_bytes((uint8_t *)&f, sizeof(float));
-        *value = f;
-    } else if (tlv->content.used == 8) {
-        memcpy(value, tlv->content.data, sizeof(double));
-        swap_bytes((uint8_t *)value, sizeof(double));
-    } else
-        return -EINVAL;
-
-    SOL_DBG("TLV has float data. Value: %g", *value);
-    return 0;
-}
-
-SOL_API int
-sol_lwm2m_tlv_get_obj_link(struct sol_lwm2m_tlv *tlv,
-    uint16_t *object_id, uint16_t *instance_id)
-{
-    int32_t i = 0;
-
-    SOL_NULL_CHECK(tlv, -EINVAL);
-    SOL_NULL_CHECK(object_id, -EINVAL);
-    SOL_NULL_CHECK(instance_id, -EINVAL);
-    SOL_INT_CHECK(is_resource(tlv), < 0, -EINVAL);
-    LWM2M_TLV_CHECK_API(tlv, -EINVAL);
-    SOL_INT_CHECK(tlv->content.used, != OBJ_LINK_LEN, -EINVAL);
-
-
-    memcpy(&i, tlv->content.data, OBJ_LINK_LEN);
-    swap_bytes((uint8_t *)&i, OBJ_LINK_LEN);
-    *object_id = (i >> 16) & 0xFFFF;
-    *instance_id = i & 0xFFFF;
-
-    SOL_DBG("TLV has object link value. Object id:%" PRIu16
-        "  Instance id:%" PRIu16 "", *object_id, *instance_id);
-    return 0;
-}
-
-SOL_API int
-sol_lwm2m_tlv_get_bytes(struct sol_lwm2m_tlv *tlv, struct sol_buffer *buf)
-{
-    SOL_NULL_CHECK(tlv, -EINVAL);
-    SOL_NULL_CHECK(buf, -EINVAL);
-    SOL_INT_CHECK(is_resource(tlv), < 0, -EINVAL);
-    LWM2M_TLV_CHECK_API(tlv, -EINVAL);
-
-    return sol_buffer_append_bytes(buf, (uint8_t *)tlv->content.data, tlv->content.used);
-}
-
-SOL_API void
-sol_lwm2m_resource_clear(struct sol_lwm2m_resource *resource)
-{
-    uint16_t i;
-
-    SOL_NULL_CHECK(resource);
-    LWM2M_RESOURCE_CHECK_API(resource);
-
-    if (resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE ||
-        resource->data_type == SOL_LWM2M_RESOURCE_DATA_TYPE_STRING) {
-        for (i = 0; i < resource->data_len; i++)
-            sol_blob_unref(resource->data[i].content.blob);
-    }
-    free(resource->data);
-    resource->data = NULL;
 }
 
 static int
@@ -3013,6 +420,15 @@ clear_bootstrap_ctx(struct sol_lwm2m_client *client)
         client->bootstrap_ctx.timeout = NULL;
         client->bootstrap_ctx.server_uri = NULL;
     }
+}
+
+static void
+clear_resource_array(struct sol_lwm2m_resource *array, uint16_t len)
+{
+    uint16_t i;
+
+    for (i = 0; i < len; i++)
+        sol_lwm2m_resource_clear(&array[i]);
 }
 
 /* Returns 1 if authorized; 0 if unauthorized and < 0 if error */
@@ -4402,15 +1818,6 @@ err_exit:
     return r;
 }
 
-static void
-clear_resource_array(struct sol_lwm2m_resource *array, uint16_t len)
-{
-    uint16_t i;
-
-    for (i = 0; i < len; i++)
-        sol_lwm2m_resource_clear(&array[i]);
-}
-
 static int
 read_resources(struct sol_lwm2m_client *client,
     struct obj_ctx *obj_ctx, struct obj_instance *instance,
@@ -4915,161 +2322,6 @@ lifetime_client_timeout(void *data)
     if (spam_update(data, true) < 0)
         SOL_WRN("Could not spam the update");
     return false;
-}
-
-static bool
-bootstrap_reply(void *data, struct sol_coap_server *server,
-    struct sol_coap_packet *req, const struct sol_network_link_addr *cliaddr)
-{
-    struct bootstrap_ctx *ctx = data;
-    uint8_t code = 0;
-
-    if (!cliaddr && !req)
-        code = SOL_COAP_RESPONSE_CODE_GATEWAY_TIMEOUT;
-
-    if (!code)
-        sol_coap_header_get_code(req, &code);
-    ((void (*)(void *,
-    struct sol_lwm2m_bootstrap_server *,
-    struct sol_lwm2m_bootstrap_client_info *, const char *,
-    enum sol_coap_response_code))ctx->cb)
-        ((void *)ctx->data, ctx->server, ctx->cinfo, ctx->path, code);
-
-    if (code != SOL_COAP_RESPONSE_CODE_GATEWAY_TIMEOUT)
-        send_ack_if_needed(server, req, cliaddr);
-    free(ctx->path);
-    free(ctx);
-    return false;
-}
-
-static int
-send_bootstrap_packet(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    enum bootstrap_type type, void *cb, const void *data,
-    enum sol_coap_method method,
-    struct sol_lwm2m_resource *resources,
-    struct sol_lwm2m_resource **instances, size_t *instances_len,
-    uint16_t *instances_ids, size_t len)
-{
-    int r;
-    struct sol_coap_packet *pkt;
-    struct bootstrap_ctx *ctx;
-
-    r = setup_coap_packet(method, SOL_COAP_MESSAGE_TYPE_CON,
-        NULL, path, NULL, NULL, resources, instances, instances_len,
-        instances_ids, len, NULL, &pkt);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (!cb)
-        return sol_coap_send_packet(server->coap, pkt, &client->cliaddr);
-
-    ctx = malloc(sizeof(struct bootstrap_ctx));
-    SOL_NULL_CHECK_GOTO(ctx, err_exit);
-
-    ctx->path = strdup(path);
-    SOL_NULL_CHECK_GOTO(ctx->path, err_exit);
-    ctx->type = type;
-    ctx->server = server;
-    ctx->cinfo = client;
-    ctx->data = data;
-    ctx->cb = cb;
-
-    return sol_coap_send_packet_with_reply(server->coap, pkt, &client->cliaddr,
-        bootstrap_reply, ctx);
-
-err_exit:
-    free(ctx);
-    sol_coap_packet_unref(pkt);
-    return -ENOMEM;
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_write_object(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    struct sol_lwm2m_resource **instances, size_t *instances_len,
-    uint16_t *instances_ids, size_t len,
-    void (*cb)(void *data,
-    struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_bootstrap_packet(server, client, path,
-        BOOTSTRAP_WRITE, cb, data, SOL_COAP_METHOD_PUT, NULL, instances,
-        instances_len, instances_ids, len);
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_write(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    struct sol_lwm2m_resource *resources, size_t len,
-    void (*cb)(void *data,
-    struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_bootstrap_packet(server, client, path,
-        BOOTSTRAP_WRITE, cb, data, SOL_COAP_METHOD_PUT, resources, NULL, NULL, NULL, len);
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_delete_object_instance(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    void (*cb)(void *data,
-    struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client, const char *path,
-    enum sol_coap_response_code response_code),
-    const void *data)
-{
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-    SOL_NULL_CHECK(path, -EINVAL);
-
-    return send_bootstrap_packet(server, client, path,
-        BOOTSTRAP_DELETE, cb, data, SOL_COAP_METHOD_DELETE, NULL, NULL, NULL, NULL, 0);
-}
-
-SOL_API int
-sol_lwm2m_bootstrap_server_send_finish(struct sol_lwm2m_bootstrap_server *server,
-    struct sol_lwm2m_bootstrap_client_info *client)
-{
-    struct sol_coap_packet *pkt;
-    int r;
-
-    SOL_NULL_CHECK(server, -EINVAL);
-    SOL_NULL_CHECK(client, -EINVAL);
-
-    pkt = sol_coap_packet_new_request(SOL_COAP_METHOD_POST, SOL_COAP_MESSAGE_TYPE_CON);
-    r = -ENOMEM;
-    SOL_NULL_CHECK_GOTO(pkt, err_exit);
-
-    r = sol_coap_add_option(pkt, SOL_COAP_OPTION_URI_PATH, "bs", strlen("bs"));
-    SOL_INT_CHECK_GOTO(r, < 0, err_coap);
-
-    SOL_DBG("Sending Bootstrap Finish to LWM2M Client %s", client->name);
-    r = sol_coap_send_packet(server->coap, pkt, &client->cliaddr);
-
-    r = sol_ptr_vector_remove(&server->clients, client);
-    if (r < 0)
-        SOL_WRN("Could not remove the client %s from the clients list",
-            client->name);
-    bootstrap_client_info_del(client);
-
-    return r;
-
-err_coap:
-    sol_coap_packet_unref(pkt);
-err_exit:
-    return r;
 }
 
 static int
