+++ src/lib/comms/sol-lwm2m-common.h
@@ -1,0 +1,679 @@
+/*
+ * This file is part of the Soletta (TM) Project
+ *
+ * Copyright (C) 2015 Intel Corporation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <errno.h>
+#include <float.h>
+#include <stdlib.h>
+#include <time.h>
+#include <math.h>
+#include <ctype.h>
+
+#define SOL_LOG_DOMAIN &_lwm2m_domain
+
+#include "sol-log-internal.h"
+#include "sol-util-internal.h"
+#include "sol-list.h"
+#include "sol-lwm2m.h"
+#include "sol-macros.h"
+#include "sol-mainloop.h"
+#include "sol-monitors.h"
+#include "sol-random.h"
+#include "sol-str-slice.h"
+#include "sol-str-table.h"
+#include "sol-util.h"
+#include "sol-http.h"
+
+SOL_LOG_INTERNAL_DECLARE_STATIC(_lwm2m_domain, "lwm2m");
+
+#define LWM2M_BOOTSTRAP_QUERY_PARAMS (1)
+#define LWM2M_UPDATE_QUERY_PARAMS (4)
+#define LWM2M_REGISTER_QUERY_PARAMS (5)
+#define NUMBER_OF_PATH_SEGMENTS (3)
+#define DEFAULT_SHORT_SERVER_ID (0)
+#define DEFAULT_CLIENT_LIFETIME (86400)
+#define DEFAULT_BINDING_MODE (SOL_LWM2M_BINDING_MODE_U)
+#define DEFAULT_LOCATION_PATH_SIZE (10)
+#define TLV_TYPE_MASK (192)
+#define TLV_ID_SIZE_MASK (32)
+#define TLV_CONTENT_LENGTH_MASK (24)
+#define TLV_CONTENT_LENGHT_CUSTOM_MASK (7)
+#define REMOVE_SIGN_BIT_MASK (127)
+#define SIGN_BIT_MASK (128)
+#define ID_HAS_16BITS_MASK (32)
+#define OBJ_LINK_LEN (4)
+#define LEN_IS_8BITS_MASK (8)
+#define LEN_IS_16BITS_MASK (16)
+#define LEN_IS_24BITS_MASK (24)
+#define UINT24_MAX (16777215)
+#define ONE_SECOND (1000)
+
+#define SECURITY_SERVER_OBJECT_ID (0)
+#define SECURITY_SERVER_URI (0)
+#define SECURITY_SERVER_IS_BOOTSTRAP (1)
+#define SECURITY_SERVER_ID (10)
+#define SECURITY_SERVER_CLIENT_HOLD_OFF_TIME (11)
+#define SECURITY_SERVER_BOOTSTRAP_SERVER_ACCOUNT_TIMEOUT (12)
+
+#define SERVER_OBJECT_ID (1)
+#define SERVER_OBJECT_SERVER_ID (0)
+#define SERVER_OBJECT_LIFETIME (1)
+#define SERVER_OBJECT_BINDING (7)
+
+#define ACCESS_CONTROL_OBJECT_ID (2)
+#define ACCESS_CONTROL_OBJECT_OBJECT_RES_ID (0)
+#define ACCESS_CONTROL_OBJECT_INSTANCE_RES_ID (1)
+#define ACCESS_CONTROL_OBJECT_ACL_RES_ID (2)
+#define ACCESS_CONTROL_OBJECT_OWNER_RES_ID (3)
+
+#ifndef SOL_NO_API_VERSION
+#define LWM2M_TLV_CHECK_API(_tlv, ...) \
+    do { \
+        if (SOL_UNLIKELY((_tlv)->api_version != \
+            SOL_LWM2M_TLV_API_VERSION)) { \
+            SOL_WRN("Couldn't handle tlv that has unsupported version " \
+                "'%u', expected version is '%u'", \
+                (_tlv)->api_version, SOL_LWM2M_TLV_API_VERSION); \
+            return __VA_ARGS__; \
+        } \
+    } while (0);
+#define LWM2M_RESOURCE_CHECK_API(_resource, ...) \
+    do { \
+        if (SOL_UNLIKELY((_resource)->api_version != \
+            SOL_LWM2M_RESOURCE_API_VERSION)) { \
+            SOL_WRN("Couldn't handle resource that has unsupported version " \
+                "'%u', expected version is '%u'", \
+                (_resource)->api_version, SOL_LWM2M_RESOURCE_API_VERSION); \
+            return __VA_ARGS__; \
+        } \
+    } while (0);
+#define LWM2M_RESOURCE_CHECK_API_GOTO(_resource, _label) \
+    do { \
+        if (SOL_UNLIKELY((_resource).api_version != \
+            SOL_LWM2M_RESOURCE_API_VERSION)) { \
+            SOL_WRN("Couldn't handle resource that has unsupported version " \
+                "'%u', expected version is '%u'", \
+                (_resource).api_version, SOL_LWM2M_RESOURCE_API_VERSION); \
+            goto _label; \
+        } \
+    } while (0);
+#define LWM2M_OBJECT_CHECK_API(_obj, ...) \
+    do { \
+        if (SOL_UNLIKELY((_obj).api_version != \
+            SOL_LWM2M_OBJECT_API_VERSION)) { \
+            SOL_WRN("Couldn't handle object that has unsupported version " \
+                "'%u', expected version is '%u'", \
+                (_obj).api_version, SOL_LWM2M_OBJECT_API_VERSION); \
+            return __VA_ARGS__; \
+        } \
+    } while (0);
+#define LWM2M_OBJECT_CHECK_API_GOTO(_obj, _label) \
+    do { \
+        if (SOL_UNLIKELY((_obj).api_version != \
+            SOL_LWM2M_OBJECT_API_VERSION)) { \
+            SOL_WRN("Couldn't handle object that has unsupported version " \
+                "'%u', expected version is '%u'", \
+                (_obj).api_version, SOL_LWM2M_OBJECT_API_VERSION); \
+            goto _label; \
+        } \
+    } while (0);
+#else
+#define LWM2M_TLV_CHECK_API(_tlv, ...)
+#define LWM2M_RESOURCE_CHECK_API(_resource, ...)
+#define LWM2M_OBJECT_CHECK_API(_obj, ...)
+#define LWM2M_RESOURCE_CHECK_API_GOTO(_resource, _label)
+#define LWM2M_OBJECT_CHECK_API_GOTO(_obj, _label)
+#endif
+
+enum tlv_length_size_type {
+    LENGTH_SIZE_CHECK_NEXT_TWO_BITS = 0,
+    LENGTH_SIZE_8_BITS = 8,
+    LENGTH_SIZE_16_BITS = 16,
+    LENGTH_SIZE_24_BITS = 24
+};
+
+enum lwm2m_parser_args_state {
+    STATE_NEEDS_DIGIT = 0,
+    STATE_NEEDS_COMMA_OR_EQUAL = (1 << 1),
+    STATE_NEEDS_COMMA = (1 << 2),
+    STATE_NEEDS_APOSTROPHE = (1 << 3),
+    STATE_NEEDS_CHAR_OR_DIGIT = (1 << 4),
+};
+
+struct lifetime_ctx {
+    struct sol_timeout *timeout;
+    uint32_t lifetime;
+};
+
+struct sol_lwm2m_client_object {
+    struct sol_ptr_vector instances;
+    uint16_t id;
+};
+
+static void
+send_ack_if_needed(struct sol_coap_server *coap, struct sol_coap_packet *msg,
+    const struct sol_network_link_addr *cliaddr)
+{
+    struct sol_coap_packet *ack;
+    uint8_t type;
+
+    sol_coap_header_get_type(msg, &type);
+
+    if (type == SOL_COAP_MESSAGE_TYPE_CON) {
+        ack = sol_coap_packet_new(msg);
+        SOL_NULL_CHECK(ack);
+        if (sol_coap_send_packet(coap, ack, cliaddr) < 0)
+            SOL_WRN("Could not send the response ACK");
+    }
+}
+
+static enum sol_lwm2m_binding_mode
+get_binding_mode_from_str(const struct sol_str_slice binding)
+{
+    static const struct sol_str_table map[] = {
+        SOL_STR_TABLE_PTR_ITEM("U", SOL_LWM2M_BINDING_MODE_U),
+        //TODO: The modes below are not supported for now.
+        SOL_STR_TABLE_PTR_ITEM("UQ", SOL_LWM2M_BINDING_MODE_UNKNOWN),
+        SOL_STR_TABLE_PTR_ITEM("S", SOL_LWM2M_BINDING_MODE_UNKNOWN),
+        SOL_STR_TABLE_PTR_ITEM("SQ", SOL_LWM2M_BINDING_MODE_UNKNOWN),
+        SOL_STR_TABLE_PTR_ITEM("US", SOL_LWM2M_BINDING_MODE_UNKNOWN),
+        SOL_STR_TABLE_PTR_ITEM("UQS", SOL_LWM2M_BINDING_MODE_UNKNOWN),
+        { }
+    };
+
+    return sol_str_table_lookup_fallback(map, binding,
+        SOL_LWM2M_BINDING_MODE_UNKNOWN);
+}
+
+static void
+client_objects_clear(struct sol_ptr_vector *objects)
+{
+    uint16_t i, j, *id;
+    struct sol_lwm2m_client_object *object;
+
+    SOL_PTR_VECTOR_FOREACH_IDX (objects, object, i) {
+        SOL_PTR_VECTOR_FOREACH_IDX (&object->instances, id, j)
+            free(id);
+        sol_ptr_vector_clear(&object->instances);
+        free(object);
+    }
+
+    sol_ptr_vector_clear(objects);
+}
+
+static struct sol_lwm2m_client_object *
+find_client_object_by_id(struct sol_ptr_vector *objects,
+    uint16_t id)
+{
+    uint16_t i;
+    struct sol_lwm2m_client_object *cobject;
+
+    SOL_PTR_VECTOR_FOREACH_IDX (objects, cobject, i) {
+        if (cobject->id == id)
+            return cobject;
+    }
+
+    return NULL;
+}
+
+static int
+add_to_monitors(struct sol_monitors *monitors, sol_monitors_cb_t cb, const void *data)
+{
+    struct sol_monitors_entry *m;
+
+    SOL_NULL_CHECK(cb, -EINVAL);
+
+    m = sol_monitors_append(monitors, cb, data);
+    SOL_NULL_CHECK(m, -ENOMEM);
+
+    return 0;
+}
+
+static int
+remove_from_monitors(struct sol_monitors *monitors, sol_monitors_cb_t cb, const void *data)
+{
+    int i;
+
+    SOL_NULL_CHECK(cb, -EINVAL);
+
+    i = sol_monitors_find(monitors, cb, data);
+    SOL_INT_CHECK(i, < 0, i);
+
+    return sol_monitors_del(monitors, i);
+}
+
+static size_t
+get_int_size(int64_t i)
+{
+    if (i >= INT8_MIN && i <= INT8_MAX)
+        return 1;
+    if (i >= INT16_MIN && i <= INT16_MAX)
+        return 2;
+    if (i >= INT32_MIN && i <= INT32_MAX)
+        return 4;
+    return 8;
+}
+
+static int
+get_resource_len(const struct sol_lwm2m_resource *resource, uint16_t index,
+    size_t *len)
+{
+    switch (resource->data_type) {
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_STRING:
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE:
+        *len = resource->data[index].content.blob->size;
+        return 0;
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_INT:
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_TIME:
+        *len = get_int_size(resource->data[index].content.integer);
+        return 0;
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL:
+        *len = 1;
+        return 0;
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT:
+        *len = 8;
+        return 0;
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK:
+        *len = OBJ_LINK_LEN;
+        return 0;
+    default:
+        return -EINVAL;
+    }
+}
+
+static void
+swap_bytes(uint8_t *to_swap, size_t len)
+{
+#if __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
+    return;
+#elif __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
+    uint8_t swap;
+    size_t i, j, stop;
+
+    stop = len / 2;
+
+    for (i = 0, j = len - 1; i != stop; i++, j--) {
+        swap = to_swap[i];
+        to_swap[i] = to_swap[j];
+        to_swap[j] = swap;
+    }
+#else
+#error "Unknown byte order"
+#endif
+}
+
+static int
+add_float_resource(struct sol_buffer *buf, double fp, size_t len)
+{
+    uint8_t *bytes = NULL;
+    float f;
+    double d;
+
+    if (len == 4) {
+        f = (float)fp;
+        swap_bytes((uint8_t *)&f, len);
+        bytes = (uint8_t *)&f;
+    } else {
+        d = fp;
+        swap_bytes((uint8_t *)&d, len);
+        bytes = (uint8_t *)&d;
+    }
+
+    return sol_buffer_append_bytes(buf, bytes, len);
+}
+
+static int
+add_int_resource(struct sol_buffer *buf, int64_t i, size_t len)
+{
+    swap_bytes((uint8_t *)&i, len);
+    return sol_buffer_append_bytes(buf, (uint8_t *)&i, len);
+}
+
+static int
+add_resource_bytes_to_buffer(const struct sol_lwm2m_resource *resource,
+    struct sol_buffer *buf, uint16_t idx)
+{
+    int r;
+    uint8_t b;
+    size_t len;
+
+    r = get_resource_len(resource, idx, &len);
+    SOL_INT_CHECK(r, < 0, r);
+
+    switch (resource->data_type) {
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_STRING:
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_OPAQUE:
+        return sol_buffer_append_slice(buf, sol_str_slice_from_blob(resource->data[idx].content.blob));
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_INT:
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_TIME:
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_OBJ_LINK:
+        return add_int_resource(buf, resource->data[idx].content.integer, len);
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_BOOL:
+        b = resource->data[idx].content.integer != 0 ? 1 : 0;
+        return sol_buffer_append_bytes(buf, (uint8_t *)&b, 1);
+    case SOL_LWM2M_RESOURCE_DATA_TYPE_FLOAT:
+        return add_float_resource(buf, resource->data[idx].content.fp, len);
+    default:
+        return -EINVAL;
+    }
+}
+
+static int
+set_packet_payload(struct sol_coap_packet *pkt,
+    const uint8_t *data, uint16_t len)
+{
+    struct sol_buffer *buf;
+    int r;
+
+    r = sol_coap_packet_get_payload(pkt, &buf, NULL);
+    SOL_INT_CHECK(r, < 0, r);
+
+    return sol_buffer_append_bytes(buf, data, len);
+}
+
+static int
+setup_tlv_header(enum sol_lwm2m_tlv_type tlv_type, uint16_t res_id,
+    struct sol_buffer *buf, size_t data_len)
+{
+    int r;
+    uint8_t tlv_data[6];
+    size_t tlv_data_len;
+
+    tlv_data_len = 2;
+
+    tlv_data[0] = tlv_type;
+
+    if (res_id > UINT8_MAX) {
+        tlv_data[0] |= ID_HAS_16BITS_MASK;
+        tlv_data[1] = (res_id >> 8) & 255;
+        tlv_data[2] = res_id & 255;
+        tlv_data_len++;
+    } else
+        tlv_data[1] = res_id;
+
+    if (data_len <= 7)
+        tlv_data[0] |= data_len;
+    else if (data_len <= UINT8_MAX) {
+        tlv_data[tlv_data_len++] = data_len;
+        tlv_data[0] |= LEN_IS_8BITS_MASK;
+    } else if (data_len <= UINT16_MAX) {
+        tlv_data[tlv_data_len++] = (data_len >> 8) & 255;
+        tlv_data[tlv_data_len++] = data_len & 255;
+        tlv_data[0] |= LEN_IS_16BITS_MASK;
+    } else if (data_len <= UINT24_MAX) {
+        tlv_data[tlv_data_len++] = (data_len >> 16) & 255;
+        tlv_data[tlv_data_len++] = (data_len >> 8) & 255;
+        tlv_data[tlv_data_len++] = data_len & 255;
+        tlv_data[0] |= LEN_IS_24BITS_MASK;
+    } else
+        return -ENOMEM;
+
+    r = sol_buffer_append_bytes(buf, tlv_data, tlv_data_len);
+    SOL_INT_CHECK(r, < 0, r);
+
+    return 0;
+}
+
+static int
+setup_tlv(struct sol_lwm2m_resource *resource, struct sol_buffer *buf)
+{
+    int r;
+    size_t data_len, len;
+    uint16_t i;
+    enum sol_lwm2m_tlv_type type;
+
+    LWM2M_RESOURCE_CHECK_API(resource, -EINVAL);
+
+    for (i = 0, data_len = 0; i < resource->data_len; i++) {
+        r = get_resource_len(resource, i, &len);
+        SOL_INT_CHECK(r, < 0, r);
+        data_len += len;
+    }
+
+    switch (resource->type) {
+    case SOL_LWM2M_RESOURCE_TYPE_SINGLE:
+        type = SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE;
+        break;
+    case SOL_LWM2M_RESOURCE_TYPE_MULTIPLE:
+        type = SOL_LWM2M_TLV_TYPE_MULTIPLE_RESOURCES;
+        data_len += (resource->data_len * 2);
+        break;
+    default:
+        SOL_WRN("Unknown resource type '%d'", (int)resource->type);
+        return -EINVAL;
+    }
+
+    r = setup_tlv_header(type, resource->id, buf, data_len);
+    SOL_INT_CHECK(r, < 0, r);
+
+    if (type == SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE)
+        return add_resource_bytes_to_buffer(resource, buf, 0);
+
+    for (i = 0; i < resource->data_len; i++) {
+        r = get_resource_len(resource, i, &data_len);
+        SOL_INT_CHECK(r, < 0, r);
+        r = setup_tlv_header(SOL_LWM2M_TLV_TYPE_RESOURCE_INSTANCE, resource->data[i].id,
+            buf, data_len);
+        SOL_INT_CHECK(r, < 0, r);
+        r = add_resource_bytes_to_buffer(resource, buf, i);
+        SOL_INT_CHECK(r, < 0, r);
+    }
+
+    return 0;
+}
+
+static int
+resources_to_tlv(struct sol_lwm2m_resource *resources,
+    size_t len, struct sol_buffer *tlvs)
+{
+    int r;
+    size_t i;
+
+    for (i = 0; i < len; i++) {
+        r = setup_tlv(&resources[i], tlvs);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    return 0;
+
+exit:
+    return r;
+}
+
+static int
+instances_to_tlv(struct sol_lwm2m_resource **instances,
+    size_t *instances_len, uint16_t *instances_ids, size_t len, struct sol_buffer *tlvs)
+{
+    int r;
+    size_t i, j, res, instance_data_len, data_len;
+
+    for (i = 0; i < len; i++) {
+        for (res = 0, instance_data_len = 0; res < instances_len[i]; res++)
+            for (j = 0; j < instances[i][res].data_len; j++) {
+                r = get_resource_len(&instances[i][res], j, &data_len);
+                SOL_INT_CHECK(r, < 0, r);
+
+                instance_data_len += data_len;
+            }
+
+        r = setup_tlv_header(SOL_LWM2M_TLV_TYPE_OBJECT_INSTANCE,
+            instances_ids[i], tlvs, instance_data_len);
+        SOL_INT_CHECK(r, < 0, r);
+
+        r = resources_to_tlv(instances[i], instances_len[i], tlvs);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    return 0;
+
+exit:
+    return r;
+}
+
+static int
+add_coap_int_option(struct sol_coap_packet *pkt,
+    enum sol_coap_option opt, const void *data, uint16_t len)
+{
+    uint8_t buf[sizeof(int64_t)] = { };
+
+    memcpy(buf, data, len);
+    swap_bytes(buf, len);
+    return sol_coap_add_option(pkt, opt, buf, len);
+}
+
+static int
+get_coap_int_option(struct sol_coap_packet *pkt,
+    enum sol_coap_option opt, uint16_t *value)
+{
+    const void *v;
+    uint16_t len;
+
+    v = sol_coap_find_first_option(pkt, opt, &len);
+
+    if (!v)
+        return -ENOENT;
+
+    memcpy(value, v, len);
+    swap_bytes((uint8_t *)value, len);
+    return 0;
+}
+
+static int
+setup_coap_packet(enum sol_coap_method method,
+    enum sol_coap_message_type type, const char *objects_path, const char *path,
+    uint8_t *obs, int64_t *token, struct sol_lwm2m_resource *resources,
+    struct sol_lwm2m_resource **instances, size_t *instances_len,
+    uint16_t *instances_ids, size_t len,
+    const char *execute_args,
+    struct sol_coap_packet **pkt)
+{
+    struct sol_buffer buf = SOL_BUFFER_INIT_EMPTY;
+    struct sol_buffer tlvs =
+        SOL_BUFFER_INIT_FLAGS(NULL, 0, SOL_BUFFER_FLAGS_NO_NUL_BYTE);
+    struct sol_random *random;
+    uint16_t content_type, content_len = 0;
+    const uint8_t *content_data;
+    int64_t t;
+    int r;
+
+    random = sol_random_new(SOL_RANDOM_DEFAULT, 0);
+    SOL_NULL_CHECK(random, -ENOMEM);
+
+    *pkt = sol_coap_packet_new_request(method, type);
+    r = -ENOMEM;
+    SOL_NULL_CHECK_GOTO(*pkt, exit);
+
+    r = sol_random_get_int64(random, &t);
+    if (r < 0) {
+        SOL_WRN("Could not generate a random number");
+        goto exit;
+    }
+
+    r = sol_coap_header_set_token(*pkt, (uint8_t *)&t,
+        (uint8_t)sizeof(int64_t));
+    if (r < 0) {
+        SOL_WRN("Could not set the token");
+        goto exit;
+    }
+    SOL_DBG("Setting token as %" PRId64 ", len = %zu", t, sizeof(int64_t));
+
+    if (token)
+        *token = t;
+
+    if (obs) {
+        r = add_coap_int_option(*pkt, SOL_COAP_OPTION_OBSERVE, obs,
+            sizeof(uint8_t));
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    if (objects_path) {
+        r = sol_buffer_append_slice(&buf,
+            sol_str_slice_from_str(objects_path));
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    r = sol_buffer_append_slice(&buf, sol_str_slice_from_str(path));
+    SOL_INT_CHECK_GOTO(r, < 0, exit);
+
+    if (strcmp(path, "/") != 0) {
+        r = sol_coap_packet_add_uri_path_option(*pkt, buf.data);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    if (execute_args) {
+        size_t str_len;
+        content_type = SOL_LWM2M_CONTENT_TYPE_TEXT;
+        content_data = (const uint8_t *)execute_args;
+        str_len = strlen(execute_args);
+        r = -ENOMEM;
+        SOL_INT_CHECK_GOTO(str_len, >= UINT16_MAX, exit);
+        content_len = str_len;
+    } else if (resources) {
+        content_type = SOL_LWM2M_CONTENT_TYPE_TLV;
+        r = resources_to_tlv(resources, len, &tlvs);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+        r = -ENOMEM;
+        SOL_INT_CHECK_GOTO(tlvs.used, >= UINT16_MAX, exit);
+        content_data = tlvs.data;
+        content_len = tlvs.used;
+    } else if (instances) {
+        content_type = SOL_LWM2M_CONTENT_TYPE_TLV;
+        r = instances_to_tlv(instances, instances_len, instances_ids, len, &tlvs);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+        r = -ENOMEM;
+        SOL_INT_CHECK_GOTO(tlvs.used, >= UINT16_MAX, exit);
+        content_data = tlvs.data;
+        content_len = tlvs.used;
+    }
+
+    if (content_len > 0) {
+        r = add_coap_int_option(*pkt, SOL_COAP_OPTION_CONTENT_FORMAT,
+            &content_type, sizeof(content_type));
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+
+        r = set_packet_payload(*pkt, content_data, content_len);
+        SOL_INT_CHECK_GOTO(r, < 0, exit);
+    }
+
+    r = 0;
+
+exit:
+    if (r < 0)
+        sol_coap_packet_unref(*pkt);
+    sol_buffer_fini(&tlvs);
+    sol_buffer_fini(&buf);
+    sol_random_del(random);
+    return r;
+}
+
+static void
+tlv_clear(struct sol_lwm2m_tlv *tlv)
+{
+    LWM2M_TLV_CHECK_API(tlv);
+    sol_buffer_fini(&tlv->content);
+}
+
+static int
+is_resource(struct sol_lwm2m_tlv *tlv)
+{
+    if (tlv->type != SOL_LWM2M_TLV_TYPE_RESOURCE_WITH_VALUE &&
+        tlv->type != SOL_LWM2M_TLV_TYPE_RESOURCE_INSTANCE)
+        return -EINVAL;
+    return 0;
+}
