+++ src/samples/coap/lwm2m-client.c
@@ -32,8 +32,7 @@
    server received in the bootstrap information.
  */
 
-
-#include "sol-lwm2m.h"
+#include "sol-lwm2m-client.h"
 #include "sol-mainloop.h"
 #include "sol-vector.h"
 #include "sol-util.h"
